% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-appendix-subsection-refs.R
\name{add_appendix_subsection_refs}
\alias{add_appendix_subsection_refs}
\title{Post process appendix subsections in LaTeX code}
\usage{
add_appendix_subsection_refs(x)
}
\arguments{
\item{x}{The LaTeX code as a vector of character strings, one for each line}
}
\value{
The modified LaTeX code as a vector of character strings, one for
each line
}
\description{
A post-processor to modify the appendix subsection LaTeX code
after the document has been knit. The modifications ensure correct
referencing of the appendices throughout the document. References are
injected into the LaTeX code in \code{x}. The LaTeX macros injected prior to
each appendix/section/subsection/subsubsection are 'newcounter' and
'refstepcounter'`.
}
\keyword{internal}
