% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_extra_header}
\alias{add_extra_header}
\title{Adds an extra header to the top of a \code{\link[=csas_table]{csas_table()}}. Works for longpages.}
\usage{
add_extra_header(kable_input, header = NULL, bold = FALSE,
  italic = FALSE, monospace = FALSE, underline = FALSE,
  strikeout = FALSE, align = "c", color = NULL, background,
  font_size, angle, escape, line = TRUE, line_sep = 3)
}
\arguments{
\item{kable_input}{An R object, typically a matrix or data frame.}

\item{header}{a vector of character strings to use for the extra header names}

\item{bold}{See \code{\link[kableExtra:::pdfTable_add_header_above]{kableExtra::::pdfTable_add_header_above()}}}

\item{italic}{See \code{\link[kableExtra:::pdfTable_add_header_above]{kableExtra::::pdfTable_add_header_above()}}}

\item{monospace}{See \code{\link[kableExtra:::pdfTable_add_header_above]{kableExtra::::pdfTable_add_header_above()}}}

\item{underline}{See \code{\link[kableExtra:::pdfTable_add_header_above]{kableExtra::::pdfTable_add_header_above()}}}

\item{strikeout}{See \code{\link[kableExtra:::pdfTable_add_header_above]{kableExtra::::pdfTable_add_header_above()}}}

\item{align}{See \code{\link[kableExtra:::pdfTable_add_header_above]{kableExtra::::pdfTable_add_header_above()}}}

\item{color}{See \code{\link[kableExtra:::pdfTable_add_header_above]{kableExtra::::pdfTable_add_header_above()}}}

\item{background}{See \code{\link[kableExtra:::pdfTable_add_header_above]{kableExtra::::pdfTable_add_header_above()}}}

\item{font_size}{See \code{\link[kableExtra:::pdfTable_add_header_above]{kableExtra::::pdfTable_add_header_above()}}}

\item{angle}{See \code{\link[kableExtra:::pdfTable_add_header_above]{kableExtra::::pdfTable_add_header_above()}}}

\item{escape}{See \code{\link[kableExtra:::pdfTable_add_header_above]{kableExtra::::pdfTable_add_header_above()}}}

\item{line}{See \code{\link[kableExtra:::pdfTable_add_header_above]{kableExtra::::pdfTable_add_header_above()}}}

\item{line_sep}{See \code{\link[kableExtra:::pdfTable_add_header_above]{kableExtra::::pdfTable_add_header_above()}}}
}
\value{
See \code{\link[kableExtra:::pdfTable_add_header_above]{kableExtra::::pdfTable_add_header_above()}}
}
\description{
Adds an extra header to the top of a \code{\link[=csas_table]{csas_table()}}. Works for longpages.
}
