% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_resdoc_docx_frontmatter}
\alias{add_resdoc_docx_frontmatter}
\title{Add front matter to Res Doc docx file}
\usage{
add_resdoc_docx_frontmatter(
  frontmatter = "templates/RES2024-eng-frontmatter.docx",
  resdoc = "_book/resdoc-english.docx"
)
}
\arguments{
\item{frontmatter}{Path to title page file included with resdoc template}

\item{resdoc}{Path to content generated using resdoc_word}
}
\value{
A merged .docx
}
\description{
Add title page and table of contents to a Res Doc. Must hand edit
\code{templates/RES2024-eng-frontmatter.docx}to have your desired title and authors etc.
}
