% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_resdoc_docx_titlepage}
\alias{add_resdoc_docx_titlepage}
\title{Add a Res Doc titlepage to a docx file}
\usage{
add_resdoc_docx_titlepage(
  titlepage = "templates/RES2024-eng-titlepage.docx",
  resdoc = "_book/resdoc-english.docx"
)
}
\arguments{
\item{titlepage}{Filename}

\item{resdoc}{Filename}
}
\value{
A merged .docx
}
\description{
Add a Res Doc titlepage. Must hand edit \code{templates/RES2024-eng-titlepage.docx}
to have your desired title and authors etc.
}
