% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_resdoc_titlepage}
\alias{add_resdoc_titlepage}
\title{Add Res Doc titlepage}
\usage{
add_resdoc_titlepage(titlepage = "templates/RES2016-eng-titlepage.docx",
  resdoc = "_book/resdoc.docx")
}
\arguments{
\item{titlepage}{Filename}

\item{resdoc}{Filename}
}
\value{
A merged .docx
}
\description{
Add Res Doc titlepage
}
