% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_techreport_docx_titlepage}
\alias{add_techreport_docx_titlepage}
\title{Add a titlepage to a Tech report docx file}
\usage{
add_techreport_docx_titlepage(
  titlepage = ifelse(fr(), "templates/tech-report-cover-fra.docx",
    "templates/tech-report-cover-eng.docx"),
  doc = "_book/techreport.docx"
)
}
\arguments{
\item{titlepage}{Filename}

\item{doc}{Filename}
}
\value{
A merged .docx
}
\description{
Must hand edit the first two pages of your file afterwards to have your desired title and authors.
}
