% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alert}
\alias{alert}
\title{Issue a warning}
\usage{
alert(...)
}
\arguments{
\item{...}{Arguments that make up the message}
}
\description{
Issue a warning
}
\details{
Uses \link[cli:symbol]{cli::symbol} and \code{\link[crayon:crayon]{crayon::yellow()}}
}
\keyword{internal}
