% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{b}
\alias{b}
\title{Wrapper for the \verb{[base::cat()]} function so there is no separator}
\usage{
b(...)
}
\arguments{
\item{...}{\R objects (see \sQuote{Details} for the types of objects
    allowed).}
}
\description{
Wrapper for the \verb{[base::cat()]} function so there is no separator
}
