% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bail}
\alias{bail}
\title{Abort code execution}
\usage{
bail(...)
}
\arguments{
\item{...}{Arguments that make up the message}
}
\description{
Abort code execution
}
\details{
Uses \link[cli:symbol]{cli::symbol} and \code{\link[crayon:crayon]{crayon::red()}}
}
\keyword{internal}
