% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_log}
\alias{capture_log}
\title{Capture multiple log messages from a function call}
\usage{
capture_log(f)
}
\arguments{
\item{f}{The function call}
}
\value{
A function which you call like this (if you name sit \code{x}): \code{x(1)}
to extract the messages
}
\description{
Capture multiple log messages from a function call
}
\details{
Capture all messages from a function.
Use like this:
\code{x <- capture_log(function_call(args))}
\code{j <- x(1)}
\code{j <- j$logs}
\code{messages <- map_chr(j, ~{.x$message})}
}
\keyword{internal}
