% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat}
\alias{cat}
\title{Redefinition of \code{\link[=cat]{cat()}} with separator set to empty string}
\usage{
cat(..., file = "", sep = "", fill = FALSE, labels = NULL, append = FALSE)
}
\description{
Redefinition of \code{\link[=cat]{cat()}} with separator set to empty string
}
