% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catize}
\alias{catize}
\title{Reformat string to convert inline code to cat-like strings}
\usage{
catize(str, verbose = FALSE)
}
\arguments{
\item{str}{The string containing inline knitr-style R code
(backtick-enclosed)}

\item{verbose}{Shows two vectors extracted from the input text, which are
used to build the new string:
\enumerate{
\item the text chunks between the R code chunks
\item the R code chunks between the text chunks
}}
}
\value{
A non-quoted (see \code{\link[=noquote]{noquote()}} string) which can be
enclosed with double-quotes and copy/pasted into a \code{\link[=cat]{cat()}}
or \code{\link[=paste]{paste()}} command
}
\description{
Reformats the supplied string so that inline code chunks
are replaced with the format used in commands such as \code{\link[=cat]{cat()}}
and \code{\link[=paste]{paste()}} which is a series of comma-separated strings
and R objects. Use this function to reformat strings from
rmarkdown to something that can be copy/pasted into a
\code{\link[=b]{b()}} command inside a knitr chunk.

This is used primarily to convert sections of code that have been written
in rmarkdown into simple strings with r code embedded between quoted,
comma-separated strings. Here is a simple example:
Original rmarkdown string:

The date is \code{`r Sys.Date()`} today

Modified cat-like string:

The date is ", Sys.Date(), " today

\strong{This is not meant to be used in an Rmd file. You must convert your
rmarkdown paragraphs by running them one at a time through this function
outside your document build, and pasting the contents into a call to \code{\link[=b]{b()}}
inside your Rmd file.}
}
\note{

}
