% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_notify}
\alias{check_notify}
\title{Issue a check-mark notification}
\usage{
check_notify(...)
}
\arguments{
\item{...}{Arguments that make up the message}
}
\description{
Issue a check-mark notification
}
\details{
Uses \link[cli:symbol]{cli::symbol} and \code{\link[crayon:crayon]{crayon::green()}}
}
\keyword{internal}
