% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-header.R
\name{chunk_header_add}
\alias{chunk_header_add}
\title{Add an option to a knitr chunk header}
\usage{
chunk_header_add(header = NULL, option = NULL, ovr = FALSE)
}
\arguments{
\item{header}{The knitr chunk header string}

\item{option}{The option to add. This whole string will be added}

\item{ovr}{Logical. If \code{TRUE}, and \code{option} is already defined in
\code{header}, \code{option} will replace that definition completely}
}
\value{
The modified knitr chunk header string
}
\description{
Add an option to a knitr chunk header
}
\details{
Here is an example of how the option is added. If \code{option} is
"needs_trans = TRUE" then the text in \code{option} will be added with a
preceeding comma and space, at the end of the list of options that is
already present:
\verb{\{r example, eval = !fr()\}` would become }{r example, eval = !fr(), needs_trans = TRUE}\code{If}header\verb{is not matched by the regular expression used by knitr for beginning chunks [knitr::all_patterns$md$chunk.begin] then a warning will be issued and}header\verb{will be returned. If}option\verb{already exists in}header\code{and}ovr\code{is}TRUE\verb{, it will be overwritten with the new }option` value
}
\seealso{
Other knitr-chunk-header: 
\code{\link{chunk_header_contains}()},
\code{\link{chunk_header_remove}()}
}
\concept{knitr-chunk-header}
\keyword{internal}
