% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-header.R
\name{chunk_header_contains}
\alias{chunk_header_contains}
\title{Verify if a knitr chunk header contains a given option}
\usage{
chunk_header_contains(header, option)
}
\arguments{
\item{header}{The knitr chunk header string}

\item{option}{The option to check for}
}
\value{
Logical. \code{TRUE} if \code{option} was found in \code{header}. \code{FALSE}
if \code{option} was not found in \code{header} or \code{option} was \code{NULL} or \code{NA}
}
\description{
Verify if a knitr chunk header contains a given option
}
\details{
Will return \code{TRUE} if you use the wrong type of quotes, for example
if you request \code{option}  as \code{"results = 'asis'"} and the header contains
\verb{"results = \\"asis\\""}.
Spaces between equals sign are unimportant, the regular expression will
allow for any amount. User can provide any amount or none in \code{option}
and it will match any amount or none in the \code{header}.
}
\seealso{
Other knitr-chunk-header: 
\code{\link{chunk_header_add}()},
\code{\link{chunk_header_remove}()}
}
\concept{knitr-chunk-header}
\keyword{internal}
