% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable-extra-non-exports.R
\name{cline_gen}
\alias{cline_gen}
\title{standardize_header_input s a \link[kableExtra:kableExtra-package]{kableExtra::kableExtra} non-exported function}
\usage{
cline_gen(header_df, booktabs, line_sep)
}
\arguments{
\item{header_df}{header_df}

\item{booktabs}{booktabs}

\item{line_sep}{line_sep}
}
\description{
Included to pass R CMD check cleanly
}
\references{
\url{https://github.com/haozhu233/kableExtra}
}
\keyword{internal}
