% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{color_factory}
\alias{color_factory}
\title{Create functions for coloring various keyword families}
\usage{
color_factory(message, r_color)
}
\arguments{
\item{message}{The message to color}

\item{r_color}{The R color string which will be accepted by
\code{\link[crayon:make_style]{crayon::make_style()}}}
}
\value{
A function to custom color text
}
\description{
Create functions for coloring various keyword families
}
