% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv-blank-lines.R
\name{conv_blank_lines}
\alias{conv_blank_lines}
\title{Convert blank lines in Rmd code to have WYSIWYG newlines
(WYSIWYG = What You See Is What You Get)}
\usage{
conv_blank_lines(chunk)
}
\arguments{
\item{chunk}{The Rmd chunk to process}
}
\value{
A list of two elements, 1) The corrected part of the chunk and
2) the rest of the chunk starting with the line after the last blank line
}
\description{
Convert blank lines in Rmd code to WYSIWYG newlines. The provided chunk
will start with a blank line and#' possibly more. The series of blank
lines will be converted into a mini-chunk, which will be areturned
as the first element of a two-element list, the second of which is the
rest of the Rmd.
}
\examples{
library(csasdown)
chunk <- c("", "", "", "Some text..", "More Rmd text")
tmp <- conv_blank_lines(chunk)
blank_line_chunk <- tmp[[1]]
the_rest <- tmp[[2]]
}
