% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv-header-lines.R
\name{conv_header_lines}
\alias{conv_header_lines}
\title{Convert header lines in Rmd code to have WYSIWYG newlines
(WYSIWYG = What You See Is What You Get)}
\usage{
conv_header_lines(chunk)
}
\arguments{
\item{chunk}{A vector of character strings representing lines for RMD code}
}
\value{
A list of two elements, 1) The corrected part of the chunk and
2) the rest of the chunk starting with the line after the last blank line
}
\description{
Convert header lines in Rmd code to WYSIWYG newlines. The provided chunk
will start with a header line and possibly more or return NULL as the
converted chunk, and the whole \code{chunk} as the rest. The series of header
lines will be converted into a mini-chunk, which will be returned
as the first element of a two-element list, the second element is the
rest of the Rmd.
}
\examples{
library(csasdown)
chunk <- c("#Main", "##Introduction", "", "Some text..", "More Rmd text")
tmp <- conv_header_lines(chunk)
header_line_chunk <- tmp[[1]]
the_rest <- tmp[[2]]
}
\keyword{internal}
