% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv-table-lines.R
\name{conv_table_lines}
\alias{conv_table_lines}
\title{Convert table lines in Rmd code to have WYSIWYG newlines
(WYSIWYG = What You See Is What You Get)}
\usage{
conv_table_lines(chunk)
}
\arguments{
\item{chunk}{The Rmd chunk to process}
}
\value{
A list of two elements, 1) The corrected part of the chunk and
2) the rest of the chunk starting with the line after the last list line
}
\description{
Convert table lines in Rmd code to WYSIWYG newlines. There are two types
of manually-entered Rmarkdown table supported:
\enumerate{
\item The provided chunk will have a beginning and end row of dashes,
and a row of dashes after the table header. All these are directly beside
the text in a vertical sense. The table rows not directly beside the top
and bottom lines must have a blank line between them
\item The provided chunk will have a single dashed line, after the table header.
This line must have text above and below directly, no newlines. the table
rows must be right next to each other, no blank lines between them
If there is no table detected, NULL will be returned as the converted chunk,
and the whole \code{chunk} as \code{the_rest}. The series of list lines will be
converted into a mini-chunk, which will be returned as the first element of
a two-element list, the second element is the rest of the Rmd.
}
}
