% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv-type-1-table-lines.R
\name{conv_type_1_table_lines}
\alias{conv_type_1_table_lines}
\title{Convert an Rmarkdown type 1 table into a simplified form with WYSIWYG
newlines}
\usage{
conv_type_1_table_lines(chunk)
}
\arguments{
\item{chunk}{A vector of character strings representing lines for RMD code}
}
\value{
A list of two elements, which are the modified type 1 table and the
rest of the chunk starting with the line after the end of the table
}
\description{
Convert an Rmarkdown type 1 table into a simplified form with WYSIWYG
newlines. If there is no table detected, NULL will be returned as the
converted chunk, or an error will be thrown depending on the situation.
}
\details{
A type 1 table is defined as a manually-entered Rmarkdown table with a
minimum of five lines with this format:

\if{html}{\out{<div class="sourceCode">}}\preformatted{----- ----- -----
 a     b     c
----- ----- -----
abc   def   ghi
----- ----- -----
}\if{html}{\out{</div>}}

There is no whitespace in between any of these lines. The first text row is
the column headers for the table and the second text row is the table data.
The second row can be multiple rows separated by an arbitrary number of
blank lines, but there cannot be blank lines before the table data or after.
Here is an example with three table data rows in acceptable format, along
with the optional table caption text which must start with 'Table:'

\if{html}{\out{<div class="sourceCode">}}\preformatted{----- ----- -----
 a     b     c
----- ----- -----
abc   def   ghi

jkl   mno   pqr


stu   vwx   yza
----- ----- -----
Table: Table caption (0 or more blank lines between table and this caption)
A second line of table caption here (no blank lines in between)
}\if{html}{\out{</div>}}
}
\examples{
chunk <- c("---------- -----------", "  Parameter   Value",
           "---------- -----------", "     x          1.0",
           "     y          2.2", "---------- -----------")
tmp <- csasdown:::conv_type_1_table_lines(chunk)
the_rest <- tmp[[2]]
}
\keyword{internal}
