% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv-type-2-table-lines.R
\name{conv_type_2_table_lines}
\alias{conv_type_2_table_lines}
\title{Convert an Rmarkdown type 2 table into a simplified form with WYSIWYG
newlines}
\usage{
conv_type_2_table_lines(chunk)
}
\arguments{
\item{chunk}{A vector of character strings representing lines for RMD code}
}
\value{
A list of two elements, which are the modified type 1 table and the
rest of the chunk starting with the line after the end of the table
}
\description{
Convert an Rmarkdown type 2 table into a simplified form with WYSIWYG
newlines. If there is no table detected, NULL will be returned as the
converted chunk, or an error will be thrown depending on the situation.
}
\details{
A type 2 table is defined as a manually-entered Rmarkdown table with a
minimum of three lines with this format:
}
\section{a     b     c}{
#----- ----- -----

#----- ----- -----
}

\section{abc   def   ghi}{
There is no whitespace in between any of these lines. The first text row is
the column headers for the table and the second text row is the table data.
The second row can be multiple rows which must not be separated by blank
lines. They all must be together without whitespace in between.
Here is an example with three table data rows in acceptable format, along
with the optional table caption text which must start with 'Table:'


}

\section{jkl   mno   pqr}{

}

\section{stu   vwx   yza}{
#Table: Table caption (0 or more blank lines between table and this caption)
#A second line of table caption here (no blank lines in between)
}

\examples{
library(csasdown)
chunk <- c("  Parameter   Value", "---------- -----------",
           "     x          1.0", "     y          2.2")
tmp <- conv_type_2_table_lines(chunk)
the_rest <- tmp[[2]]
}
