% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-newlines-rmd.R
\name{convert_newlines_rmd}
\alias{convert_newlines_rmd}
\title{Fix newlines in text that is from inside \code{\link[=cat]{cat()}} in an Rmd document}
\usage{
convert_newlines_rmd(text_chunk)
}
\arguments{
\item{text_chunk}{A vector of strings, one line of Rmarkdown per
element}
}
\value{
A modified vector of strings
}
\description{
A vector of lines of text strings made up of Rmarkdown code, 1 line per
vector element will contain some newlines which are represented in the
vector as empty strings (""). These are replaced in a somewhat complex
way so that the worry of how newlines in Rmarkdown works is removed.
This allows for 'What you see is what you get' code in the text strings.
}
\examples{
convert_newlines_rmd(c("hello world", "", "", "Welcome!"))
}
