% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-pandoc-theme-files.R
\name{create_pandoc_theme_files}
\alias{create_pandoc_theme_files}
\title{Creates package JSON and LaTeX files for pandoc highlight theme types}
\usage{
create_pandoc_theme_files(pandoc_dir = NULL)
}
\arguments{
\item{pandoc_dir}{A directory name where pandoc command line is available.
If \code{NULL}, the environment variable \code{RSTUDIO_PANDOC} will be used}
}
\description{
Creates package JSON and LATEX files for pandoc highlight theme types.
Runs command line pandoc to extract the JSON code, then a modified function
to change the format to LaTeX
}
\details{
This is not testable on GitHub actions, so do not write tests for this
function
}
\seealso{
Other highlight-theme: 
\code{\link{gen_latex_highlight_code}()},
\code{\link{parse_pandoc_highlight_theme}()}
}
\concept{highlight-theme}
\keyword{internal}
