% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-rstudio-project-file.R
\name{create_rstudio_project_file}
\alias{create_rstudio_project_file}
\title{Creates a blank Rstudio project file in the current directory with the
same name as the current directory.}
\usage{
create_rstudio_project_file(directory = getwd(), verbose = TRUE)
}
\arguments{
\item{directory}{The directory in which to create the RStudio project file}

\item{verbose}{Logical. If \code{TRUE}, print messages}
}
\description{
Creates a blank Rstudio project file in the current directory with the
same name as the current directory.
}
\seealso{
Other drafting: 
\code{\link{draft}()}
}
\concept{drafting}
\keyword{internal}
