% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-tmp-yaml-file.R
\name{create_tmp_yaml_file}
\alias{create_tmp_yaml_file}
\title{Create a copy of a bookdown YAML file containing modified Rmd File names}
\usage{
create_tmp_yaml_file(yaml_fn = "_bookdown.yml")
}
\arguments{
\item{yaml_fn}{The \link{bookdown} YAML file name, by default is "_bookdown.yml"}
}
\value{
The name of the temporary YAML file created
}
\description{
Creates a copy of a bookdown YAML file with "tmp_" prepended to the name.
Inside the new file, all listed Rmd files will have their names prepended
with "tmp_".
It is required that all filenames appear on their own line in the file.
}
\details{
The file must have an entry like this:

\verb{rmd_files: ["index.Rmd",}
\verb{            "01-chap1.Rmd"]}
\verb{           #"02-chap2.Rmd",}
\verb{           #"03-chap3.Rmd",}
\verb{           #"04-references.Rmd",}
\verb{           #"05-appendix.Rmd"]}

This will appear in the new file as:

\verb{rmd_files: ["index.Rmd",}
\verb{            "tmp_01-chap1.Rmd"]}

The file must also have a line like this before the above list, which is
used as an index for where to insert the modified file list after:
book_filename: "resdoc"
}
