% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-tmp-yaml-rmd-files.R
\name{create_tmp_yaml_rmd_files}
\alias{create_tmp_yaml_rmd_files}
\title{Create a copy of a bookdown YAML file containing modified Rmd File names
Create a copy of the RMd files listed, with the names listed in the new
YAML file}
\usage{
create_tmp_yaml_rmd_files(yaml_fn = "_bookdown.yml", verbose = FALSE)
}
\arguments{
\item{yaml_fn}{The \link[bookdown:bookdown-package]{bookdown::bookdown} YAML file name, by default is "_bookdown.yml"}

\item{verbose}{Logical. If \code{TRUE}, print messages}
}
\value{
A list of two, 1) The name of the temporary YAML file created
2) A vector of the temporary Rmd files created
}
\description{
Creates a copy of a bookdown YAML file with "tmp_" prepended to the name.
Inside the new file, all listed Rmd files will have their names prepended
with "tmp_". Those files will also be created by copying the ones listed
in the original YAML file.
}
\details{
The file must have an entry like this:

\verb{rmd_files: ["index.Rmd",}
\verb{            "01-chap1.Rmd"]}
\verb{           #"02-chap2.Rmd",}
\verb{           #"03-chap3.Rmd",}
\verb{           #"04-references.Rmd",}
\verb{           #"05-appendix.Rmd"]}

This will appear in the new file as:

\verb{rmd_files: ["index.Rmd",}
\verb{            "tmp_01-chap1.Rmd"]}

The file must also have a line like this before the above list, which is
used as an index for where to insert the modified file list after:
book_filename: "resdoc"

There are only two requirements to how the files are listed in the
\verb{rmd_files: [} section: 1) they must begin right after the \code{[}
and 2) the closing \verb{]} must be at the end of its line with no more code
following. They look nice as listed above but could be more than one to
a line, or all on one line.

[\verb{and 2) the closing}]: R:\%60\%0Aand\%202)\%20the\%20closing\%20\%60
}
\keyword{internal}
