% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{csas_color}
\alias{csas_color}
\title{csasdown items (render type, format, chunk names, etc) color
for messages throughout the project}
\usage{
csas_color(message)
}
\value{
The customized message modified by the \code{\link[crayon:make_style]{crayon::make_style()}}
color function
}
\description{
csasdown items (render type, format, chunk names, etc) color
for messages throughout the project
}
