% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resdoc_pdf}
\alias{resdoc_pdf}
\alias{sr_pdf}
\alias{techreport_pdf}
\title{Creates an R Markdown PDF with CSAS formatting}
\usage{
resdoc_pdf(toc = TRUE, toc_depth = 3, highlight = "default",
  latex_engine = "pdflatex", french = FALSE, ...)

sr_pdf(latex_engine = "pdflatex", french = FALSE, ...)

techreport_pdf(french = FALSE, latex_engine = "pdflatex", ...)
}
\arguments{
\item{toc}{A Boolean (\code{TRUE} or \code{FALSE}) specifying whether table of contents
should be created.}

\item{toc_depth}{A positive integer.}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{latex_engine}{LaTeX engine.}

\item{french}{Logical for French (vs. English).}

\item{...}{other arguments to \code{\link[bookdown:pdf_book]{bookdown::pdf_book()}}.}
}
\value{
A modified \code{pdf_document} based on the CSAS LaTeX template.
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify using the CSAS LaTeX template and cls files.
}
\examples{
\dontrun{
 output: csasdown::resdoc_pdf
}
}
