% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csasdown.R
\docType{package}
\name{csasdown}
\alias{csasdown}
\alias{csasdown-package}
\title{csasdown: A package for creating CSAS Research Document with R Markdown / bookdown}
\description{
csasdown: A package for creating CSAS Research Document with R Markdown / bookdown
}
