% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable-extra-non-exports.R
\name{escape_latex2}
\alias{escape_latex2}
\title{escape_latex2 is a \link[kableExtra:kableExtra-package]{kableExtra::kableExtra} non-exported function}
\usage{
escape_latex2(x)
}
\arguments{
\item{x}{x}
}
\description{
Included to pass R CMD check cleanly
}
\references{
\url{https://github.com/haozhu233/kableExtra}
}
\keyword{internal}
