% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rmd_table_label.R
\name{extract_rmd_table_label}
\alias{extract_rmd_table_label}
\title{Search chunk to find an Rmarkdown table label}
\usage{
extract_rmd_table_label(chunk)
}
\arguments{
\item{chunk}{A vector of character strings representing lines for RMD code}
}
\value{
A list of length two. The elements are:
\enumerate{
\item A vector representing the lines of the label chunk or \code{NULL} if none found
\item A vector representing the remainder of the chunk after the label or
\code{NULL} if the label reached to the end of the chunk
}
}
\description{
Search chunk to find an Rmarkdown table label. There are two possible
starting lines to the table caption
}
\details{
The two caption possibilities are of the forms:
\enumerate{
\item Table: Caption text
Optional text..
More optional text..
...
\item Table: -Any amount of whitespace-
Caption text..
Optional text..
More optional text..
...
}
}
\keyword{internal}
