% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable-extra-non-exports.R
\name{ez_rep}
\alias{ez_rep}
\title{ez_rep is a \link[kableExtra:kableExtra-package]{kableExtra::kableExtra} non-exported function}
\usage{
ez_rep(x, n)
}
\arguments{
\item{x}{x}

\item{n}{n}
}
\description{
Included to pass R CMD check cleanly
}
\references{
\url{https://github.com/haozhu233/kableExtra}
}
\keyword{internal}
