% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_resdoc.R
\name{files_to_process}
\alias{files_to_process}
\title{Fetch vector of Rmd filenames from a YAML file}
\usage{
files_to_process(yaml_fn = "_bookdown.yml")
}
\arguments{
\item{yaml_fn}{The YAML file name}
}
\value{
A vector of the Rmd file names
}
\description{
Assumes files each appear on their own line in the YAML file. Example:
rmd_files: \link[="index.Rmd",
            "01-chap1.Rmd"]{"index.Rmd", "01-chap1.Rmd"}
#"02-chap2.Rmd",
#"03-chap3.Rmd",
#"04-references.Rmd",
#"05-appendix.Rmd"]

Any files that are commented out will not be included. In the above example,
the vector returned will be c("index.Rmd", "01-chap1.Rmd")
}
