% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-envs.R
\name{fix_envs}
\alias{fix_envs}
\title{Fix TEX environment values}
\usage{
fix_envs(
  x,
  include_abstract = TRUE,
  prepub = FALSE,
  highlight = "tango",
  include_section_nums = TRUE,
  fix_ref_section_name = FALSE
)
}
\arguments{
\item{x}{The LaTeX code as a vector of character strings, one for each line}

\item{include_abstract}{Logical. If \code{TRUE} include an abstract page}

\item{prepub}{Logical. If \code{TRUE} the document if pre-publication. Currently
only implemented for SR documents}

\item{highlight}{The code highlight type. Must be one of "pygments", "tango",
"espresso", "zenburn", "kate", "monochrome", "breezedark", "haddock"}

\item{include_section_nums}{Logical. If \code{TRUE}, Add section numbers to the
document. If \code{FALSE} the sections will not be numbered}

\item{fix_ref_section_name}{Logical. If \code{TRUE}, the REFERENCES section
will be renamed to French version}
}
\value{
The modified LaTeX code as a vector of character strings, one for
each line
}
\description{
A post-processor to modify LaTeX code after the document has been knit
}
\keyword{internal}
