% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn_color}
\alias{fn_color}
\title{Filename color for messages throughout the project}
\usage{
fn_color(message)
}
\value{
The customized message modified by the \code{\link[crayon:make_style]{crayon::make_style()}}
color function
}
\description{
Filename color for messages throughout the project
}
