% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-timeseries.R
\name{fsar_plot_base}
\alias{fsar_plot_base}
\title{Generate 4-panel figure using simulated fisheries data in base R}
\usage{
fsar_plot_base(in.df, language = c("English", "French"))
}
\arguments{
\item{language}{French or English}

\item{timeseries}{the timeseries to plot}
}
\value{
A graphics device with the four-panel of figures
}
\description{
Generate 4-panel figure using simulated fisheries data in base R
}
\examples{
fsar_plot_base(sim_fsar_data(format="long"))

}
