% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-highlight-latex.R
\name{gen_latex_highlight_code}
\alias{gen_latex_highlight_code}
\title{Generate LaTeX code for pandoc highlight-style list}
\usage{
gen_latex_highlight_code(json_lst)
}
\arguments{
\item{json_lst}{A list of pandoc highlight-style JSON codes, one for each
theme to make a latex version for}
}
\value{
A list of vectors of lines of latex code representing the highlight-style for pandoc
for each theme type
}
\description{
Generate LaTeX code for pandoc highlight-style list
}
\seealso{
Other highlight-theme: 
\code{\link{create_pandoc_theme_files}()},
\code{\link{parse_pandoc_highlight_theme}()}
}
\concept{highlight-theme}
\keyword{internal}
