% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-book-filename.R
\name{get_book_filename}
\alias{get_book_filename}
\title{Extract the book filename from the YAML file (usually index.Rmd)}
\usage{
get_book_filename(yaml_fn = "_bookdown.yml")
}
\arguments{
\item{yaml_fn}{The YAML file name. The default is '_bookdown.yml' for}
}
\value{
The book filename (typically index.Rmd) for \link{bookdown}
}
\description{
Extract the book filename from the YAML file (usually index.Rmd)
}
\details{
\code{book_file} is the "skeleton.Rmd" file, which is typically called
"index.Rmd" by default in a project. It is found in the bookdown YAML file
}
