% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-contact-info.R
\name{get_contact_info}
\alias{get_contact_info}
\title{Get regional CSAS contact information}
\usage{
get_contact_info(region = "National Capital Region")
}
\arguments{
\item{region}{Region in which the document is published; character vector.
(i.e., Pacific). Default is "National Capital Region."}
}
\value{
Email address and mailing address as list of character vectors
}
\description{
Get regional CSAS contact information including email address and mailing
address for the last page in the section "This report is available from
the." Contact information for the national CSAS office is returned if
regional information is not found, with a warning.
}
