% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-doc-type.R
\name{get_doc_type}
\alias{get_doc_type}
\title{Get the document type, either 'pdf' or 'word'}
\usage{
get_doc_type(fn = "index.Rmd")
}
\arguments{
\item{fn}{The name of the YAML file, typically 'index.Rmd' for bookdown}
}
\value{
A character string representing the document type
}
\description{
Get the document type, either 'pdf' or 'word'
}
