% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_email}
\alias{get_email}
\title{Return regional CSAS email address for the last page in the section "This
report is available from the."}
\usage{
get_email(region = "National Capital Region", isFrench = FALSE)
}
\arguments{
\item{region}{Region in which the document is published; character vector
(i.e., Pacific). Default is "National Capital Region."}

\item{isFrench}{Logical (default FALSE). Is the report in French or not?}
}
\value{
Email address as a character vector
}
\description{
Return regional CSAS email address for the last page in the section "This
report is available from the."
}
