% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-french.R
\name{get_french}
\alias{get_french}
\title{Get French YAML tag value in the index.Rmd file ('true' or 'false')}
\usage{
get_french(fn = "index.Rmd")
}
\arguments{
\item{fn}{The name of the YAML file, typically 'index.Rmd' for bookdown}
}
\value{
\code{TRUE} if the value is 'true' or \code{FALSE} if the value is 'false'
}
\description{
Get French YAML tag value in the index.Rmd file ('true' or 'false')
}
