% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-index-filename.R
\name{get_index_filename}
\alias{get_index_filename}
\title{Extract the index filename from the YAML file (usually index.Rmd)}
\usage{
get_index_filename(yaml_fn = "_bookdown.yml", verbose = FALSE)
}
\arguments{
\item{yaml_fn}{The YAML file name. The default is '_bookdown.yml' for}

\item{verbose}{Logical. If \code{TRUE}, print messages}
}
\value{
The book filename (typically index.Rmd) for \link[bookdown:bookdown-package]{bookdown::bookdown}
}
\description{
Extract the index filename from the YAML file (usually index.Rmd)
}
\details{
\code{index_file} is the code in the "skeleton.Rmd" file, which is called
'index.Rmd' by default in a project. It is found in the bookdown YAML file,
\code{yaml_fn}
}
\keyword{internal}
