% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-real-line-num.R
\name{get_real_line_num}
\alias{get_real_line_num}
\title{Get the line number in the original source file for the first R code
line (non-comment line) in a chunk}
\usage{
get_real_line_num(header, line_offsets)
}
\arguments{
\item{header}{A knitr chunk header that matches the regex
\code{all_patterns$md$chunk.begin}}

\item{line_offsets}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing the
line offsets to be added to line numbers in messages in order to give the
correct line numbers for the original files. Has columns \code{fn} (chr),
\code{chunk_header} (chr)  \code{chunk_ind} (int), \code{pre_num} (dbl), \code{post_num} (dbl),
and \code{rmd_num} (dbl). See \code{return} values for \code{\link[=inject_rmd_files]{inject_rmd_files()}} and
\code{\link[=remove_comments_from_chunks]{remove_comments_from_chunks()}}}
}
\value{
The line number in the original file
}
\description{
Get the line number in the original source file for the first R code
line (non-comment line) in a chunk
}
\details{
If \code{\link[=remove_comments_from_chunks]{remove_comments_from_chunks()}} was run and had a \code{pre_num} (in
\code{line_offsets}) value for the chunk that matched the \code{header},
that will be added to the original \code{chunk_ind} (in \code{line_offsets})
value and any additions from \code{\link[=inject_rmd_files]{inject_rmd_files()}} will be subtracted
(\code{rmd_num} in \code{line_offsets}).
}
\keyword{internal}
