% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-highlight-latex.R
\name{hex2rgb}
\alias{hex2rgb}
\title{Convert a hex string into a vector of three decimal values (RGB)}
\usage{
hex2rgb(hex, rel = FALSE, ret_alpha = FALSE)
}
\arguments{
\item{hex}{The hex string of 6 or 8 digits (if alpha included). May of may not begin with  #}

\item{rel}{If \code{TRUE}, divide the RGB values by 255 for relative values}

\item{ret_alpha}{if \code{TRUE} alpha value will be included in the output vector as the last item, so it will be
length 4 instead of 3}
}
\value{
A vector of three RGB decimal values, or three relative values
}
\description{
Convert a hex string into a vector of three decimal values (RGB)
}
\keyword{internal}
