% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject-rmd-files.R
\name{inject_rmd_files}
\alias{inject_rmd_files}
\title{Inject verbatim text from Rmd files into Rmd code}
\usage{
inject_rmd_files(rmd_files, line_offsets = NULL, verbose = FALSE)
}
\arguments{
\item{rmd_files}{A vector of character strings representing the names
of Rmd files
\code{\link[=remove_comments_from_chunks]{remove_comments_from_chunks()}}. If \code{NULL}, it will be ignored. If
present, the values will be added/subtracted from line number messages
in this function to give an accurate line number for the original files}

\item{line_offsets}{A \code{\link[=data.frame]{data.frame()}} or \code{\link[tibble:tibble]{tibble::tibble()}} containing the
line offsets to be added to line numbers in messages in order to give the
correct line numbers for the original files. Has columns \code{fn} (chr),
\code{chunk_header} (chr)  \code{chunk_ind} (int), \code{pre_num} (dbl), \code{post_num} (dbl),
and \code{rmd_num} (dbl). See \code{return} values for \code{\link[=inject_rmd_files]{inject_rmd_files()}} and
\code{\link[=remove_comments_from_chunks]{remove_comments_from_chunks()}}}

\item{verbose}{Logical. If \code{TRUE}, print messages}
}
\value{
Nothing
}
\description{
Inject verbatim text from Rmd files into Rmd code wrapped in
\code{\link[=cat]{cat()}} so that the bilingual features of csasdown can be used properly
but the author can write paragraphs or sections in pure Rmarkdown
}
\keyword{internal}
