% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject-vec-in-vec.R
\name{inject_vec_in_vec}
\alias{inject_vec_in_vec}
\title{Inject a vector of strings into another at given indices, replacing the
elements at those indices}
\usage{
inject_vec_in_vec(main_vec, sec_vec, inds)
}
\arguments{
\item{main_vec}{The vector to insert into}

\item{sec_vec}{The vector to be inserted in one or more places}

\item{inds}{The indices at which to insert the vectors}
}
\value{
A vector of strings
}
\description{
Inject a vector of strings into another at given indices, replacing the
elements at those indices, creating a new, longer vector.
}
\examples{
library(csasdown)
a <- letters[1:10]
b <- LETTERS[25:26]
inject_vec_in_vec(a, b, c(2, 5, 8))
}
\keyword{internal}
