% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable-extra-non-exports.R
\name{input_escape}
\alias{input_escape}
\title{input_escape is a \link[kableExtra:kableExtra-package]{kableExtra::kableExtra} non-exported function}
\usage{
input_escape(x, latex_align)
}
\arguments{
\item{x}{x}

\item{latex_align}{latex_align}
}
\description{
Included to pass R CMD check cleanly
}
\references{
\url{https://github.com/haozhu233/kableExtra}
}
\keyword{internal}
