% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_rmd_dashed_line}
\alias{is_rmd_dashed_line}
\title{Checks to see if character strings are dashed lines}
\usage{
is_rmd_dashed_line(lines)
}
\arguments{
\item{lines}{The vector of character strings to check}
}
\value{
A logical vector representing whether or not the lines are
dashed lines
}
\description{
Checks to see if character strings are dashed lines of any length,
including length 1
}
\keyword{internal}
