% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_rmd_header_line}
\alias{is_rmd_header_line}
\title{Checks to see if character strings are Rmarkdown header lines}
\usage{
is_rmd_header_line(lines)
}
\arguments{
\item{lines}{The vector of character strings to check}
}
\value{
A logical vector representing whether or not the lines are
Rmarkdown header lines
}
\description{
Checks to see if character strings are Rmarkdown header lines
}
\details{
A header line must be indented less than 4 spaces and start with a #
followed by one or more spaces, and the n text
}
