% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_rmd_table_line}
\alias{is_rmd_table_line}
\title{Checks to see if character strings represent the start of a Rmarkdown
tables}
\usage{
is_rmd_table_line(lines_lst)
}
\arguments{
\item{lines_lst}{A list of character strings vectors of at least length 5
for a type 1 table and 3 for a type 2 table}
}
\value{
A character vector representing which Rmarkdown table type each
element in \code{lines_lst} is
}
\description{
Checks to see if character strings represent the start of a Rmarkdown
tables
}
\details{
Three lines from the beginning of the table are required to determine if a
table is possibly valid
}
