% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_rmd_text_line}
\alias{is_rmd_text_line}
\title{Checks to see if character strings are Rmarkdown header lines}
\usage{
is_rmd_text_line(lines)
}
\arguments{
\item{lines}{The vector of character strings to check}
}
\value{
A logical vector representing whether or not the lines are
Rmarkdown header lines, which are normal text
}
\description{
Checks to see if character strings are Rmarkdown header lines which are
text with at least one character, possibly containing whitespace. In other
words, typical paragraph or sentence text with punctuation.
}
\details{
The one thing that is not a text line is a header line (Starts with a #
followed by a space)
}
