% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable-extra-non-exports.R
\name{latex_color}
\alias{latex_color}
\title{latex_color is a \link[kableExtra:kableExtra-package]{kableExtra::kableExtra} non-exported function}
\usage{
latex_color(colors, escape = TRUE)
}
\arguments{
\item{colors}{colors}

\item{escape}{escape}
}
\description{
Included to pass R CMD check cleanly
}
\references{
\url{https://github.com/haozhu233/kableExtra}
}
\keyword{internal}
