% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable-extra-non-exports.R
\name{latex_color_}
\alias{latex_color_}
\title{latex_color_ is a \link[kableExtra:kableExtra-package]{kableExtra::kableExtra} non-exported function}
\usage{
latex_color_(color)
}
\arguments{
\item{color}{color}
}
\description{
Included to pass R CMD check cleanly
}
\references{
\url{https://github.com/haozhu233/kableExtra}
}
\keyword{internal}
