% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{notify}
\alias{notify}
\title{Issue a notification}
\usage{
notify(...)
}
\arguments{
\item{...}{Arguments that make up the message}
}
\description{
Issue a notification
}
\details{
Uses \link[cli:symbol]{cli::symbol} and \code{\link[crayon:crayon]{crayon::green()}}
}
\keyword{internal}
