% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc-curr-ver-is-before.R
\name{pandoc_curr_ver_is_before}
\alias{pandoc_curr_ver_is_before}
\title{Checks whether or not the current version of Pandoc installed is below a
given version}
\usage{
pandoc_curr_ver_is_before(ver = "3.1.8")
}
\arguments{
\item{ver}{The version to compare the current version against. If \code{NULL},
an error is thrown}
}
\value{
If the current version of Pandoc is prior to this dotted string
version, \code{TRUE}. If not \code{FALSE} is returned.
}
\description{
Checks whether or not the current version of Pandoc installed is below a
given version
}
