% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-highlight-latex.R
\name{parse_pandoc_highlight_theme}
\alias{parse_pandoc_highlight_theme}
\title{Read in pandoc theme file code as a vector of lines of code
and parse it to create a list of theme tokens}
\usage{
parse_pandoc_highlight_theme(json = NULL, num_attrs = 5)
}
\arguments{
\item{json}{JSON code as a vector of lines of code extracted using a system call like this:
pandoc --print-highlight-style breezedark}

\item{num_attrs}{The number of attributes for each style token in the file.
Style tokens are labeled in the file as text-color, italics, bold, etc.}
}
\value{
A list of 1 element each for the general colors (text and background colors) and 1 element
which is a list of lists of named key-value pairs, where each list represents a chunk of attributes
as defined in the grammar above
}
\description{
Assumed Grammar of the JSON theme file based on output of:
\verb{pandoc --print-highlight-style breezedark > my_style.theme}
\itemize{
\item \code{hexval} is a six-digit quoted string representing a hex
color in this format: \code{"#ffffff"} or null if not included
\item \code{logicalval} is true or false
\verb{\{4spc\}} means 4 spaces
}

Grammar:

\if{html}{\out{<div class="sourceCode">}}\preformatted{\{
\{4spc\}"text-color": hexval,
\{4spc\}"background-color": hexval,
\{4spc\}"line-number-color": hexval
\{4spc\}"text-styles": \{
\{8spc\}"Attribute1_Name": \{
\{12spc\}"text-color": hexval,
\{12spc\}"background-color": hexval,
\{12spc\}"bold": logicalval,
\{12spc\}"italic": logicalval,
\{12spc\}"underline": logicalval
\{8spc\}\},
...
\{4spc\}"text-color": hexval,
\{4spc\}"background-color": hexval,
\{4spc\}"line-number-color": hexval
\{4spc\}"text-styles": \{
\{8spc\}"AttributeN_Name": \{
\{12spc\}"text-color": hexval,
\{12spc\}"background-color": hexval,
\{12spc\}"bold": logicalval,
\{12spc\}"italic": logicalval,
\{12spc\}"underline": logicalval
\{8spc\}\}
\{4spc\}\}
\}
}\if{html}{\out{</div>}}
}
\keyword{internal}
