% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable-extra-non-exports.R
\name{pdfTable_new_header_generator}
\alias{pdfTable_new_header_generator}
\title{pdfTable_new_header_generator is a \link[kableExtra:kableExtra-package]{kableExtra::kableExtra} non-exported function}
\usage{
pdfTable_new_header_generator(
  header_df,
  booktabs = FALSE,
  bold,
  italic,
  monospace,
  underline,
  strikeout,
  align,
  color,
  background,
  font_size,
  angle,
  line_sep,
  border_left,
  border_right
)
}
\arguments{
\item{header_df}{header_df}

\item{booktabs}{booktabs}

\item{bold}{bold}

\item{italic}{italic}

\item{monospace}{monospace}

\item{underline}{underline}

\item{strikeout}{strikeout}

\item{align}{align}

\item{color}{color}

\item{background}{background}

\item{font_size}{font_size}

\item{angle}{angle}

\item{line_sep}{line_sep}

\item{border_left}{border_left}

\item{border_right}{border_right}
}
\description{
Included to pass R CMD check cleanly
}
\references{
\url{https://github.com/haozhu233/kableExtra}
}
\keyword{internal}
