% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-chunks.R
\name{preprocess_chunks}
\alias{preprocess_chunks}
\title{Run the pre-processor on the chunks in the Rmd files}
\usage{
preprocess_chunks(fns, yaml_fn = "_bookdown.yml", verbose = FALSE)
}
\arguments{
\item{fns}{A vector of character strings holding the Rmd filenames that
are to be included}

\item{yaml_fn}{The Bookdown YAML filename}

\item{verbose}{Logical. If \code{TRUE}, print messages}
}
\value{
Nothing
}
\description{
Run the pre-processor on the chunks in the Rmd files
}
\keyword{internal}
