% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{question}
\alias{question}
\title{Issue a question notification}
\usage{
question(...)
}
\arguments{
\item{...}{Arguments that make up the message}
}
\description{
Issue a question notification
}
\details{
Uses \link[cli:symbol]{cli::symbol} and \code{\link[crayon:crayon]{crayon::green()}}
}
\keyword{internal}
