% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd-file.R
\name{read_rmd_file}
\alias{read_rmd_file}
\title{Read in am Rmd file and return it}
\usage{
read_rmd_file(fn)
}
\arguments{
\item{fn}{The Rmd filename. An extension is not needed}
}
\value{
Verbatim text representing the file
}
\description{
Read in am Rmd file and return it verbatim, with a newline added at the end
}
\details{
This allows the author to write pure Rmarkdown in a file, and have it read
into the \code{\link[=cat]{cat()}} command in a csasdown resdoc Rmd file's code chunk. See
'resdoc-bilingual' type example doc (\link{draft("resdoc-bilingual")})
}
