% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_info}
\alias{region_info}
\title{Information about national DFO region CSAS offices.}
\format{
A data frame with 7 rows and 5 variables:
\describe{
\item{Region}{English region name}
\item{RegionFr}{French region name}
\item{Email}{Email address for the region}
\item{Address}{English physical address}
\item{AddressFr}{French physical address}
}
}
\usage{
region_info
}
\description{
A dataset containing the English and French region names, English
and French addresses, and email for the CSAS offices in the region.
}
\keyword{datasets}
