% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-comments-from-chunks.R
\name{remove_comments_from_chunks}
\alias{remove_comments_from_chunks}
\title{Remove all comment lines from knitr chunks in the Rmd code for the files
provided}
\usage{
remove_comments_from_chunks(rmd_files, verbose = FALSE)
}
\arguments{
\item{rmd_files}{A vector of character strings representing the names
of Rmd files}

\item{verbose}{Logical. If \code{TRUE}, print messages}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing the Rmd filenames (column \code{fn})
and all the full chunk headers (column \code{chunk_header}) which are
unique document-wide. The chunk line number (column \code{chunk_ind}) is
also included so that headers can be searched for later, matched,
and correct line numbers from unmodified Rmd lines can be reported.
Columns containing the number of comment lines and blank lines removed
from each chunk before (\code{pre_num}) and after (\code{post_num}) a call to
\code{cat()}, \code{rmd_file()}, or a mirror chunk \verb{<<chunk-name>>} are included.
Addition of \code{chunk_ind} and \code{pre_num} allow exact line numbers to be
quoted in messages for the original files.
}
\description{
Remove all comment lines from knitr chunks in the Rmd code for the files
provided
}
\details{
Overwrites the files \code{rmd_files} with a modified version that has the
comments removed from chunks which contain a call to \code{cat()},
\code{rmd_file()}, or a mirror chunk \verb{<<chunk-name>>}
}
\keyword{internal}
