% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-html-comments.R
\name{remove_html_comments}
\alias{remove_html_comments}
\title{Remove HTML comments (Markdown comments) from a chunk of code}
\usage{
remove_html_comments(chunk, fn = "unknown")
}
\arguments{
\item{chunk}{A vector of character strings representing lines for RMD code}

\item{fn}{Filename that to reference when writing errors only. The file is
not used or looked for in any way, all the processing is done for \code{chunk}}
}
\value{
The modified vector of Rmarkdown code, which has all comments
removed and \code{NA}'s for lines which were taken up by all comments
}
\description{
Remove HTML comments (Markdown comments) from a chunk of code
}
