% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-output-files.R
\name{rename_output_files}
\alias{rename_output_files}
\title{Rename the PDF, Word, and TEX files to include English or French in the name}
\usage{
rename_output_files(index_fn, verbose = FALSE)
}
\arguments{
\item{index_fn}{The working copy of this file:
\code{csasdown/inst/rmarkdown/templates/resdoc-b/skeleton/skeleton.Rmd}
which is \code{index.Rmd} by default}
}
\value{
Nothing
}
\description{
Rename the PDF, Word, and TEX files to include English or French in the name
so that rendering in one language doesn't overwrite the other document
}
\keyword{internal}
