% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_sar}
\alias{render_sar}
\alias{render_fsar}
\title{Render a SAR/FSAR}
\usage{
render_sar(config_file = "_bookdown.yml", ...)
}
\arguments{
\item{config_file}{YAML configuration file.}

\item{...}{Arguments to pass to \code{\link[bookdown:render_book]{bookdown::render_book()}}.}
}
\description{
Render a SAR/FSAR
}
