% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-vecs-with-vecs.R
\name{replace_vecs_with_vecs}
\alias{replace_vecs_with_vecs}
\title{Inject vectors of strings into another vector replacing the elements
given the start and end indices in the main vector}
\usage{
replace_vecs_with_vecs(
  main_vec = NULL,
  repl_vecs = NULL,
  start_inds = NULL,
  end_inds = NULL
)
}
\arguments{
\item{main_vec}{The vector to insert into}

\item{repl_vecs}{A list of vectors to be inserted in one or more places}

\item{start_inds}{A vector of indices at which to insert the vectors.
This argument will have \code{sort(unique())} run on it}

\item{end_inds}{A vector of indices to mark the end of insertion.
This argument will have \code{sort(unique())} run on it}
}
\value{
A vector of strings
}
\description{
Inject vectors of strings into another vector replacing the elements
given the start and end indices in the main vector
}
\details{
Algorithm - Make a list of vector slices of the \code{main_vec} that are not
being replaced, than interleave the vectors in \code{repl_vecs} with those.
If the \code{start_inds} gives a value of 1 for the initial vector to replace,
that is a special case, and there is code to capture that and do it
correctly. The resulting list is flattened and returned.
}
\keyword{internal}
