% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resdoc_gitbook}
\alias{resdoc_gitbook}
\title{Creates an R Markdown gitbook Res Doc}
\usage{
resdoc_gitbook(...)
}
\arguments{
\item{...}{other arguments to \code{\link[bookdown:gitbook]{bookdown::gitbook()}}}
}
\value{
A gitbook webpage
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a webpage version of the resdoc.
}
