% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resdoc_pdf}
\alias{resdoc_pdf}
\title{Creates an R Markdown PDF Res Doc}
\usage{
resdoc_pdf(toc = TRUE, toc_depth = 3, highlight = "default", ...)
}
\arguments{
\item{toc}{A Boolean (TRUE or FALSE) specifying whether table of contents should be created}

\item{toc_depth}{A positive integer}

\item{highlight}{Syntax highlighting style. Supported styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", and "haddock". Pass NULL to prevent syntax highlighting.}

\item{...}{other arguments to bookdown::pdf_book}
}
\value{
A modified \code{pdf_document} based on the Reed Senior Thesis LaTeX
template
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify using the CSAS LaTeX template and cls files.
}
\examples{
\dontrun{
 output: csasdown::resdoc_pdf
}
}
