% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resdoc_word}
\alias{resdoc_word}
\title{Creates an R Markdown Word Res Doc}
\usage{
resdoc_word(...)
}
\arguments{
\item{...}{other arguments to \code{\link[bookdown:word_document2]{bookdown::word_document2()}}}
}
\value{
A Word Document based on the CSAS Res Doc template.
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a Microsoft Word version of the resdoc.
}
