% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmd_nlines}
\alias{rmd_nlines}
\title{Supply the Rmarkdown newline code for a given number of newlines}
\usage{
rmd_nlines(num_blank_lines)
}
\arguments{
\item{num_blank_lines}{A single value for the number of newlines,
or actual blank lines required}
}
\value{
A character vector containing the sequence of code necessary
to create the number of newlines required
}
\description{
Supply the Rmarkdown newline code for a given number of newlines
}
