% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pdflatex.R
\name{run_pdflatex}
\alias{run_pdflatex}
\title{Run latexmk/pdflatex on a csasdown report}
\usage{
run_pdflatex(extra_pdflatex = 1, ...)
}
\arguments{
\item{extra_pdflatex}{Extra number of times to run \code{\link[tinytex:latexmk]{tinytex::pdflatex()}}.}

\item{...}{Arguments to pass to \code{\link[tinytex:latexmk]{tinytex::latexmk()}}.}
}
\value{
The PDF in the \verb{_book} folder should be updated.
}
\description{
Run latexmk/pdflatex on a csasdown report with extra pdflatex runs.
}
\details{
This can be useful in edge cases where \code{\link[tinytex:latexmk]{tinytex::latexmk()}} doesn't
run enough times on its own.
For example \url{https://github.com/pbs-assess/csasdown/issues/151}.
}
\examples{
\dontrun{
# After running `bookdown::render_book("index.Rmd")` or knitting, run:
run_pdflatex()
}
}
