% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-highlight-latex.R
\name{save_latex_theme_code}
\alias{save_latex_theme_code}
\title{Create latex theme files for all the pandoc highlight-styles}
\usage{
save_latex_theme_code(themes = NULL, theme_path = here::here("inst", "themes"))
}
\arguments{
\item{themes}{If NULL, run function for all themes. If not NULL, apply to
the theme(s) given.}
}
\description{
Create latex theme files for all the pandoc highlight-styles
}
\details{
Source this file and run this function to create all the theme
files for the package. Once those have been created, the package has to
be re-installed, i.e. load_all("csasdown") will not access the new theme files.
}
\keyword{internal}
