% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-french.R
\name{set_french}
\alias{set_french}
\title{Set French in the index.Rmd file to 'true' or 'false'}
\usage{
set_french(
  fn = get_index_filename(system.file("rmarkdown", "templates", "resdoc", "skeleton",
    "_bookdown.yml", package = "csasdown")),
  val = TRUE
)
}
\arguments{
\item{fn}{The name of the YAML file, typically 'index.Rmd' for bookdown}

\item{val}{\code{TRUE} or \code{FALSE}}
}
\value{
Nothing, write the file \code{fn}
}
\description{
Wrapper function to set the french: YAML tag
}
