% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-yaml-tag.R
\name{set_yaml_tag}
\alias{set_yaml_tag}
\title{Set a YAML tag to a value inside the YAML header in the file given}
\usage{
set_yaml_tag(tag = NULL, val = NULL, fn = "index.Rmd")
}
\arguments{
\item{tag}{A regular expression representing the name of the tag. Usually
just the name of the tag is enough but if it occurs more than once in the
file, you will need to refine your regular expression here}

\item{val}{The new value to set \code{tag} to. Can be a vector of character
strings, in which case they will each be placed on their own line,
separated properly by YAML syntax. Do not include the OR bar (\code{|}) or
any newline characters, they will be added by the function}

\item{fn}{The name of the YAML file, typically 'index.Rmd' for bookdown}
}
\value{
Nothing
}
\description{
Set a YAML tag to a value inside the YAML header in the file given
}
