% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake-data.R
\name{sim_fsar_data}
\alias{sim_fsar_data}
\title{Simulate fisheries data to demonstrate FSAR figures}
\usage{
sim_fsar_data(format = c("wide", "long"))
}
\arguments{
\item{format}{Long or wide data frame?}
}
\value{
A data frame
}
\description{
Simulate fisheries data to demonstrate FSAR figures
}
\examples{
sim_fsar_data()
}
