% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sr_pdf}
\alias{sr_pdf}
\title{Creates an R Markdown PDF SR}
\usage{
sr_pdf(latex_engine = "pdflatex", ...)
}
\arguments{
\item{latex_engine}{LaTeX engine.}

\item{...}{other arguments to \code{\link[bookdown:pdf_book]{bookdown::pdf_book()}}.}
}
\value{
A modified \code{pdf_document} based on the CSAS LaTeX template.
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify using the CSAS LaTeX template and cls files.
}
\examples{
\dontrun{
 output: csasdown::sr_pdf
}
}
