% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable-extra-non-exports.R
\name{standardize_header_input}
\alias{standardize_header_input}
\title{standardize_header_input s a \link[kableExtra:kableExtra-package]{kableExtra::kableExtra} non-exported function}
\usage{
standardize_header_input(header)
}
\arguments{
\item{header}{header}
}
\description{
Included to pass R CMD check cleanly
}
\references{
\url{https://github.com/haozhu233/kableExtra}
}
\keyword{internal}
