% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{stk_pop}
\alias{stk_pop}
\title{Pop a value from the top of a stack}
\usage{
stk_pop(stack)
}
\arguments{
\item{stack}{A 'stack' (character vector) or \code{NULL}}
}
\value{
A list of two elements, the first is the modified stack with the top
value removed, the second is the value that was popped off. If the stack is
empty, \code{NULL} is returned
}
\description{
Pop a value from the top of a stack
}
