% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{stk_push}
\alias{stk_push}
\title{Push a character to a stack}
\usage{
stk_push(stack, val)
}
\arguments{
\item{stack}{A 'stack' (character vector). Set to \code{NULL} for an empty stack}

\item{val}{The value to push to the stack}
}
\value{
The stack with the new value on the top
}
\description{
Push a character to a stack
}
