% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{stk_size}
\alias{stk_size}
\title{Return the size of the stack}
\usage{
stk_size(stack)
}
\arguments{
\item{stack}{A 'stack' (character vector) or \code{NULL}}
}
\value{
The size of the stack (zero if the stack is empty)
}
\description{
Return the size of the stack
}
