% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tag_color}
\alias{tag_color}
\title{YAML tag color for messages throughout the project}
\usage{
tag_color(message)
}
\value{
The customized message modified by the \code{\link[crayon:make_style]{crayon::make_style()}}
color function
}
\description{
YAML tag color for messages throughout the project
}
