% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/techreport-docx.R
\name{techreport_docx}
\alias{techreport_docx}
\title{Create .docx Tech Reports}
\usage{
techreport_docx(...)
}
\arguments{
\item{...}{Other arguments to \code{\link[officedown:rdocx_document]{officedown::rdocx_document()}}}
}
\value{
A Word Document in the \code{.docx} format based on the Tech Report
template
}
\description{
This is a function called within the YAML of the
\code{index.Rmd} file to specify the creation of a .docx version of a
Tech Report
}
