% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-csasstyle.R
\name{update_csasstyle}
\alias{update_csasstyle}
\title{Copy the csas-style directory from the local library to the current
directory}
\usage{
update_csasstyle(
  copy = TRUE,
  line_nums = TRUE,
  line_nums_mod = 1,
  lot_lof = FALSE,
  draft_watermark = FALSE,
  which_sty = "res-doc.sty"
)
}
\arguments{
\item{copy}{Logical. If \code{TRUE}, copy and overwrite if the directory
already exists. If \code{FALSE}, only copy if the directory does not exist in
the current directory}

\item{line_nums}{Logical. If \code{TRUE}, include line numbering in the document}

\item{line_nums_mod}{Numerical. Which modulo line numbers to label,
e.g. 2 = every second line, 3 = every 3rd line, etc.}

\item{lot_lof}{Logical. If \code{TRUE}, include list of tables and list of
figures in the document. Implemented only for 'resdoc' and 'techreport'}

\item{draft_watermark}{Logical. If \code{TRUE}, show a DRAFT watermark on all
pages of the output document}

\item{which_sty}{Name of the style file to modify}
}
\value{
Nothing
}
\description{
Copy the csas-style directory from the local library location to
the current directory, overwriting and editing the style file if
necessary
}
\keyword{internal}
