% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_chunk_headers.R
\name{validate_chunk_headers}
\alias{validate_chunk_headers}
\title{Validate that chunk headers contain only certain combinations of settings}
\usage{
validate_chunk_headers(
  rmd_files,
  en_chunk_regex = "^\\\\S+-en$",
  fr_chunk_regex = "^\\\\S+-fr$",
  verbose = FALSE
)
}
\arguments{
\item{rmd_files}{A vector of character strings representing the names
of Rmd files}

\item{en_chunk_regex}{A regular expression to match for the chunk
name for English chunks. Default is 'ends in -en'. The \code{$} means anchor
to the end, so '-en' must be at the end. \verb{\\\\S+} means match one or more
any other non-whitespace characters}

\item{fr_chunk_regex}{A regular expression to match for the chunk
name for French chunks. Default is 'ends in -fr'}

\item{verbose}{Logical. If \code{TRUE}, print messages}
}
\value{
A vector of error messages, one item for each chunk problem
}
\description{
Validate that chunk headers contain only certain combinations of settings.
Check that chunks with \code{eval = fr()} or \code{eval = !fr()}, have the correct
\code{needs_trans} setting and a correct chunk name format if \code{chunk_regex}
is \code{TRUE}. Verify that these chunks also contain the \code{results = 'asis'}
header option. Without that, \code{cat()} produces non-Rmarkdown output
in the document.
}
\details{
Inserts \code{results = 'asis'} if necessary for chunks that have the
\code{eval = fr()} or \code{eval = !fr()} header options.
}
\keyword{internal}
