% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_index_file.R
\name{validate_index_file}
\alias{validate_index_file}
\title{Validate the input file for required YAML entries for a
specific document type}
\usage{
validate_index_file(fn = "index.Rmd", doc_type)
}
\arguments{
\item{fn}{The bookdown Rmd file which contains the YAML headers required
to run \code{\link[bookdown:render_book]{bookdown::render_book()}}}

\item{doc_type}{The type of document to validate. Can be one of 'resdoc',
'sr', 'techreport' or any of those with appended underscore and more text,
for example 'resdoc_pdf' or 'sr_word'}
}
\value{
\code{TRUE} or \code{FALSE} for file validation
}
\description{
Validate the input file for required YAML entries for a
specific document type
}
\details{
Warnings will be given for missing items prior to returning \code{FALSE} for
missing YAML tags in the file \code{fn}
}
