% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_appendix_subsection_refs}
\alias{add_appendix_subsection_refs}
\title{Fix the appendix subsections so they can be referenced properly in text}
\usage{
add_appendix_subsection_refs(x)
}
\arguments{
\item{x}{A vector of lines of the TEX file}
}
\value{
Modified TEX lines ( a vector of lines of the TEX file)
}
\description{
Fix the appendix subsections so they can be referenced properly in text
}
