% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_extra_header}
\alias{add_extra_header}
\title{Adds an extra header to the top of a \code{\link[=csas_table]{csas_table()}}. Works for longpages.}
\usage{
add_extra_header(
  kable_input,
  header = NULL,
  bold = FALSE,
  italic = FALSE,
  monospace = FALSE,
  underline = FALSE,
  strikeout = FALSE,
  align = "c",
  color = NULL,
  background,
  font_size,
  angle,
  escape,
  line = TRUE,
  line_sep = 3
)
}
\arguments{
\item{kable_input}{An R object, typically a matrix or data frame.}

\item{header}{a vector of character strings to use for the extra header names}

\item{bold}{See kableExtra:::pdfTable_add_header_above()}

\item{italic}{See kableExtra:::pdfTable_add_header_above()}

\item{monospace}{See kableExtra:::pdfTable_add_header_above()}

\item{underline}{See kableExtra:::pdfTable_add_header_above()}

\item{strikeout}{See kableExtra:::pdfTable_add_header_above()}

\item{align}{See kableExtra:::pdfTable_add_header_above()}

\item{color}{See kableExtra:::pdfTable_add_header_above()}

\item{background}{See kableExtra:::pdfTable_add_header_above()}

\item{font_size}{See kableExtra:::pdfTable_add_header_above()}

\item{angle}{See kableExtra:::pdfTable_add_header_above()}

\item{escape}{See kableExtra:::pdfTable_add_header_above()}

\item{line}{See kableExtra:::pdfTable_add_header_above()}

\item{line_sep}{See kableExtra:::pdfTable_add_header_above()}
}
\value{
See kableExtra:::pdfTable_add_header_above()
}
\description{
Adds an extra header to the top of a \code{\link[=csas_table]{csas_table()}}. Works for longpages.
}
