% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{b}
\alias{b}
\title{Redefines the \verb{[base::cat()]} function so there is no separator}
\usage{
b(..., file = "", sep = "", fill = FALSE, labels = NULL, append = FALSE)
}
\arguments{
\item{...}{\R objects (see \sQuote{Details} for the types of objects
    allowed).}

\item{file}{A \link[base]{connection}, or a character string naming the file
    to print to.  If \code{""} (the default), \code{cat} prints to the
    standard output connection, the console unless redirected by
    \code{\link[base]{sink}}.
#ifdef unix not active  }

\item{sep}{a character vector of strings to append after each element.}

\item{fill}{a logical or (positive) numeric controlling how the output is
    broken into successive lines.  If \code{FALSE} (default), only newlines
    created explicitly by \samp{"\\n"} are printed.  Otherwise, the
    output is broken into lines with print width equal to the option
    \code{width} if \code{fill} is \code{TRUE}, or the value of
    \code{fill} if this is numeric.  Linefeeds are only inserted
    \emph{between} elements, strings wider than \code{fill} are not
    wrapped.  Non-positive \code{fill} values are
    ignored, with a warning.}

\item{labels}{character vector of labels for the lines printed.
    Ignored if \code{fill} is \code{FALSE}.}

\item{append}{logical. Only used if the argument \code{file} is the
    name of file (and not a connection or \code{"|cmd"}).
    If \code{TRUE} output will be appended to
    \code{file}; otherwise, it will overwrite the contents of
    \code{file}.}
}
\description{
Redefines the \verb{[base::cat()]} function so there is no separator
}
