% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat}
\alias{cat}
\title{Redefinition of \code{\link[=cat]{cat()}} with separator set to empty string}
\usage{
cat(..., file = "", sep = " ", fill = FALSE, labels = NULL, append = FALSE)
}
\arguments{
\item{...}{\code{R} objects (see \code{Details} for the types of objects allowed)}

\item{file}{A connection, or a character string naming the file to print
to. If "" (the default), cat prints to the standard output connection
the console unless redirected by sink}

\item{sep}{A character vector of strings to append after each element.}

\item{fill}{A logical or (positive) numeric controlling how the output is
broken into successive lines. If \code{FALSE} (default), only newlines created
explicitly by ‘⁠\\n"⁠’ are printed. Otherwise, the output is broken in
to lines with print width equal to the option width if fill is \code{TRUE},
or the value of \code{fill} if this is numeric. Linefeeds are only inserted
between elements, strings wider than \code{fill} are not wrapped. Non-positive
\code{fill} values are ignored, with a warning}

\item{labels}{A character vector of labels for the lines printed. Ignored
if \code{fill} is \code{FALSE}}

\item{append}{Logical. Only used if the argument file is the name of file
(and not a connection or "\verb{|cmd}"). If \code{TRUE} output will be appended to file;
otherwise, it will overwrite the contents of file.}
}
\description{
Redefinition of \code{\link[=cat]{cat()}} with separator set to empty string
}
\details{
\code{cat} is useful for producing output in user-defined functions.
  It converts its arguments to character vectors, concatenates
  them to a single character vector, appends the given \code{sep = }
  string(s) to each element and then outputs them.

  No line feeds (aka \dQuote{newline}s) are output unless explicitly
  requested by \samp{"\\n"}
  or if generated by filling (if argument \code{fill} is \code{TRUE} or
  numeric).

  If \code{file} is a connection and open for writing it is written from
  its current position.  If it is not open, it is opened for the
  duration of the call in \code{"wt"} mode and then closed again.

  Currently only \link[base]{atomic} vectors and \link[base]{name}s are handled,
  together with \code{NULL} and other zero-length objects (which produce
  no output).  Character strings are output \sQuote{as is} (unlike
  \code{\link[base]{print.default}} which escapes non-printable characters and
  backslash --- use \code{\link[base]{encodeString}} if you want to output
  encoded strings using \code{cat}).  Other types of \R object should be
  converted (e.g., by \code{\link[base]{as.character}} or \code{\link[base]{format}})
  before being passed to \code{cat}.  That includes factors, which are
  output as integer vectors.

  \code{cat} converts numeric/complex elements in the same way as
  \code{print} (and not in the same way as \code{\link[base]{as.character}}
  which is used by the S equivalent), so \code{\link[base]{options}}
  \code{"digits"} and \code{"scipen"} are relevant.  However, it uses
  the minimum field width necessary for each element, rather than the
  same field width for all elements.
}
\note{
If any element of \code{sep} contains a newline character, it is
  treated as a vector of terminators rather than separators, an element
  being output after every vector element \emph{and} a newline after the
  last.  Entries are recycled as needed.
}
\examples{
iter <- stats::rpois(1, lambda = 10)
## print an informative message
cat("iteration = ", iter <- iter + 1, "\n")

## 'fill' and label lines:
cat(paste(letters, 100* 1:26), fill = TRUE, labels = paste0("{", 1:10, "}:"))
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988)
  \emph{The New S Language}.
  Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link[base]{print}}, \code{\link[base]{format}}, and \code{\link[base]{paste}}
  which concatenates into a string.
}
\keyword{internal}
