% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catize}
\alias{catize}
\title{Reformat the supplied string so that inline code chunks
are replaced with the format used in commands such as \code{\link[=cat]{cat()}}
and \code{\link[=paste]{paste()}} which is a series of comma-separated strings
and R objects. Use this function to reformat strings from
rmarkdown to something that can be copy/pasted into a
b("") command inside a knitr chunk.}
\usage{
catize(str, verbose = FALSE)
}
\arguments{
\item{str}{The string containing inline knitr-style R code
(backtick-enclosed)}
}
\value{
A non-quoted (see \code{\link[=noquote]{noquote()}} string) which can be
enclosed with double-quotes and copy/pasted into a \code{\link[=cat]{cat()}}
or \code{\link[=paste]{paste()}} command
}
\description{
Reformat the supplied string so that inline code chunks
are replaced with the format used in commands such as \code{\link[=cat]{cat()}}
and \code{\link[=paste]{paste()}} which is a series of comma-separated strings
and R objects. Use this function to reformat strings from
rmarkdown to something that can be copy/pasted into a
b("") command inside a knitr chunk.
}
