% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-yaml.R
\name{check_yaml}
\alias{check_yaml}
\title{Check to make sure \code{index.Rmd} contains all current YAML options}
\usage{
check_yaml(
  type = c("resdoc", "resdoc_pdf", "resdoc_word", "sr", "sr_pdf", "sr_word",
<<<<<<< HEAD
    "techreport", "techreport_pdf", "techreport_word", "fsar_word"),
=======
    "manureport", "manureport_pdf", "manureport_word", "techreport", "techreport_pdf",
    "techreport_word"),
>>>>>>> 5fca72471dcb59199d5022642e2fd08cd3a452a7
  verbose = FALSE
)
}
\arguments{
\item{type}{Type of document}

\item{verbose}{Logical. If \code{TRUE}, print messages}
}
\description{
As the csasdown package is updated, sometimes new mandatory YAML options are added
to the \code{index.Rmd} file. Running this function will compare your file to the
version built into the currently installed version of csasdown and issue
am error message telling you what doesn't match if needed.
}
