% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_yaml}
\alias{check_yaml}
\title{Check to make sure index.Rmd contains all current YAML options}
\usage{
check_yaml(type = "resdoc")
}
\arguments{
\item{type}{Type of document. Currently only in summative or research
documents. Currently this is only implemented for research documents.}
}
\description{
As the packages updated sometimes new mandatory options are added to the
\code{index.Rmd} file. Running this function will compare your file to the version
built into the currently installed version of csasdown.
}
