% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-header.R
\name{chunk_header_remove}
\alias{chunk_header_remove}
\title{Remove an option and its value from a knitr chunk header}
\usage{
chunk_header_remove(header = NULL, option = NULL)
}
\arguments{
\item{header}{The knitr chunk header string}

\item{option}{The option to remove}
}
\value{
The modified knitr chunk header string
}
\description{
Remove an option and its value from a knitr chunk header
}
\seealso{
Other knitr-chunk-header: 
\code{\link{chunk_header_add}()},
\code{\link{chunk_header_contains}()}
}
\concept{knitr-chunk-header}
\keyword{internal}
