% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv-list-lines.R
\name{conv_list_lines}
\alias{conv_list_lines}
\title{Convert list lines in Rmd code to have WYSIWYG newlines
(WYSIWYG = What You See Is What You Get)}
\usage{
conv_list_lines(chunk)
}
\arguments{
\item{chunk}{A vector of character strings representing lines for RMD code}
}
\value{
A list of two elements, 1) The corrected part of the chunk and
2) the rest of the chunk starting with the line after the last list line
}
\description{
Convert list lines in Rmd code to WYSIWYG newlines. The provided chunk
will start with a list line and possibly more or return NULL as the
converted chunk, and the whole \code{chunk} as \code{the_rest}. The series of list
lines will be converted into a mini-chunk, which will be returned
as the first element of a two-element list, the second element is the
rest of the Rmd.
}
\examples{
library(csasdown)
chunk <- c("- Item 1", "- Item 2", "", "1. Item 1", "  a. Item 1a")
tmp <- conv_list_lines(chunk)
list_line_chunk <- tmp[[1]]
the_rest <- tmp[[2]]
}
\keyword{internal}
