% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv-paragraph-lines.R
\name{conv_paragraph_lines}
\alias{conv_paragraph_lines}
\title{Convert paragraph lines in Rmd code to have WYSIWYG newlines
(WYSIWYG = What You See Is What You Get)}
\usage{
conv_paragraph_lines(chunk)
}
\arguments{
\item{chunk}{The Rmd chunk to process}
}
\value{
A list of two elements, 1) The corrected part of the chunk and
2) the rest of the chunk starting with the line after the last blank line
}
\description{
Convert paragraph lines in Rmd code to WYSIWYG newlines. The provided chunk
will start with regular text (or a R inline code chunk) and possibly more
lines. The series of paragraph lines will be converted into a mini-chunk,
which will be returned as the first element of a two-element list, the
second element is the rest of the Rmd.
}
