% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-newlines-rmd.R
\name{convert_newlines_rmd}
\alias{convert_newlines_rmd}
\title{Remove all the constraints in Rmd involving newlines}
\usage{
convert_newlines_rmd(chunk, strip_blank_lines = TRUE)
}
\arguments{
\item{chunk}{A vector of strings, one line of Rmarkdown per element}

\item{strip_blank_lines}{Strip all lines composed of whitespace.
Rstudio automatically indents several spaces to the start of lines when you
press 'Enter' and it is common to ignore those when editing}
}
\value{
A modified vector of strings
}
\description{
Rmarkdown code requires several different entries to render newlines
in a final document. These are frustrating because the algorithm
is not clear on what you need for which case. For example, header lines
(lines starting with #) do not need a newline, but if you want 'N' blank
lines between the header and next text, you need 'N lines containing
only \\, without any empty lines following.
For other text, you need N + 1 lines containing only \\ with an empty line
following, before the next text occurs. List elements must not have newlines
between them.
This allows for 'What you see is what you get' in your Rmarkdown code.
}
\keyword{internal}
