% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-mirror-chunks.R
\name{copy_mirror_chunks}
\alias{copy_mirror_chunks}
\title{Copy mirrored knitr chunk code into chunks with the mirrored call}
\usage{
copy_mirror_chunks(rmd_files)
}
\arguments{
\item{rmd}{A vector of character strings representing the lines in an
Rmd file}
}
\value{
A vector of character strings representing the lines in an
Rmd file but with the mirrored calls replaced with code
}
\description{
Copy the mirrored chunk code into the chunk with the mirrored call.
knitr does this automatically but we need to do it in the pre-processing
step so we can modify the newlines in the mirrored code
}
