% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{create_rstudio_project_file}
\alias{create_rstudio_project_file}
\title{Creates a blank Rstudio project file in the current directory with the same name
as the current directory.}
\usage{
create_rstudio_project_file()
}
\description{
Creates a blank Rstudio project file in the current directory with the same name
as the current directory.
}
