% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_resdoc.R
\name{create_tmp_yaml_file}
\alias{create_tmp_yaml_file}
\title{Create a copy of the YAML file with temporary Rmd File names}
\usage{
create_tmp_yaml_file(yaml_fn = "_bookdown.yml", book_fn = "index.Rmd")
}
\arguments{
\item{yaml_fn}{The YAML file name}

\item{book_fn}{The Rmd file containing the YAML code used for rendering.
The default is 'index.Rmd' for \code{\link[bookdown:render_book]{bookdown::render_book()}}}
}
\value{
The name of the temporary YAML file created
}
\description{
Create a copy of the YAML file with temporary Rmd File names
}
