% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datareport-docx.R, R/fsar-word.R,
%   R/manureport-docx.R, R/resdoc-word.R, R/sr-word.R, R/techreport-docx.R
\name{datareport_docx}
\alias{datareport_docx}
\alias{fsar_docx}
\alias{manureport_docx}
\alias{resdoc_docx}
\alias{sr_docx}
\alias{techreport_docx}
\title{Create \code{.docx} CSAS-formatted documents}
\usage{
datareport_docx(...)

fsar_docx(...)

manureport_docx(...)

resdoc_docx(...)

sr_docx(...)

techreport_docx(...)
}
\arguments{
\item{...}{Other arguments to \code{\link[officedown:rdocx_document]{officedown::rdocx_document()}}}
}
\value{
A \code{.docx} file
}
\description{
This is a function called within the YAML of the
\code{index.Rmd} file to specify the creation of a \code{.docx} version of a
CSAS report.
}
\examples{
# These functions would normally only be specified within the YAML section of
# index.Rmd

x <- resdoc_docx()
names(x)
}
\seealso{
\code{\link[=render]{render()}}, \code{\link[=draft]{draft()}}
}
