% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsar-word.R, R/resdoc-word.R, R/sr-word.R
\name{fsar_docx}
\alias{fsar_docx}
\alias{resdoc_docx}
\alias{sr_docx}
\title{Creates an Microsoft Word CSAS-formatted document}
\usage{
fsar_docx(...)

resdoc_docx(...)

sr_docx(...)
}
\arguments{
\item{...}{Other arguments to \code{\link[officedown:rdocx_document]{officedown::rdocx_document()}}}
}
\value{
A Word Document in the \code{.docx} format based on the FSAR template.

A Word Document in the \code{.docx} format based on the CSAS Res Doc
template

A Word Document in the \code{.docx} format based on the CSAS Res Doc
template
}
\description{
This is a function called in output in the YAML of the \code{index.Rmd} file
to specify the creation of an FSAR word document.

This is a function called within the YAML of the
\code{index.Rmd} file to specify the creation of a .docx version of a
Research Document or Science Response.

This is a function called within the YAML of the
\code{index.Rmd} file to specify the creation of a .docx version of a
Research Document or Science Response.
}
