% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsar-word.R, R/resdoc-word.R, R/sr-word.R,
%   R/techreport-docx.R
\name{fsar_docx}
\alias{fsar_docx}
\alias{resdoc_docx}
\alias{sr_docx}
\alias{techreport_docx}
\title{Create .docx CSAS-formatted documents}
\usage{
fsar_docx(...)

resdoc_docx(...)

sr_docx(...)

techreport_docx(...)
}
\arguments{
\item{...}{Other arguments to \code{\link[officedown:rdocx_document]{officedown::rdocx_document()}}}
}
\value{
A \code{.docx} file
}
\description{
This is a function called within the YAML of the
\code{index.Rmd} file to specify the creation of a .docx version of a
CSAS report.
}
