% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsar-word.R, R/resdoc-word.R
\name{fsar_word}
\alias{fsar_word}
\alias{resdoc_docx}
\title{Creates an Microsoft Word CSAS-formatted document}
\usage{
fsar_word(...)

resdoc_docx(...)
}
\arguments{
\item{...}{Other arguments to \code{\link[officedown:rdocx_document]{officedown::rdocx_document()}}}
}
\value{
A Word Document in the \code{.docx} format based on the FSAR template.

A Word Document in the \code{.docx} format based on the CSAS Res Doc template
}
\description{
This is a function called in output in the YAML of the \code{index.Rmd} file
to specify the creation of an FSAR word document.

This is a function called in output in the YAML of the \code{index.Rmd} file
to specify the creation of a Microsoft Word version of the Research
Document or Science Response.
}
