% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resdoc_word}
\alias{resdoc_word}
\alias{sr_word}
\alias{techreport_word}
\title{Creates an R Markdown Word CSAS-formatted document}
\usage{
resdoc_word(french = FALSE, ...)

sr_word(french = FALSE, ...)

techreport_word(french = FALSE, ...)
}
\arguments{
\item{french}{Logical for French (vs. English).}

\item{...}{other arguments to \code{\link[bookdown:word_document2]{bookdown::word_document2()}}}
}
\value{
A Word Document based on the CSAS Res Doc template.
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a Microsoft Word version of the Research
Document or Science Response.
}
