% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resdoc-word.R, R/sr-word.R, R/techreport-word.R
\name{resdoc_word}
\alias{resdoc_word}
\alias{sr_word}
\alias{techreport_word}
\title{Creates an Microsoft Word CSAS-formatted document}
\usage{
resdoc_word(...)

sr_word(...)

techreport_word(...)
}
\arguments{
\item{...}{Other arguments to \code{\link[bookdown:html_document2]{bookdown::word_document2()}}}
}
\value{
A Word Document in the \code{.docx} format based on the CSAS Res Doc
template
}
\description{
This is a function called in output in the YAML of the \code{index.Rmd} file
to specify the creation of a Microsoft Word version of the Research
Document or Science Response.
}
