% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resdoc-pdf.R, R/sr-pdf.R, R/techreport-pdf.R
\name{resdoc_pdf}
\alias{resdoc_pdf}
\alias{sr_pdf}
\alias{techreport_pdf}
\title{Creates an R Markdown PDF with CSAS formatting}
\usage{
resdoc_pdf(
  toc = TRUE,
  toc_depth = 3,
  highlight = "tango",
  latex_engine = "pdflatex",
  prepub = FALSE,
  copy_sty = TRUE,
  line_nums = FALSE,
  line_nums_mod = 1,
  lot_lof = FALSE,
  draft_watermark = FALSE,
  french = FALSE,
  include_section_nums = TRUE,
  pandoc_args = c("--top-level-division=chapter", "--wrap=none",
    "--default-image-extension=png"),
  ...
)

sr_pdf(
  latex_engine = "pdflatex",
  prepub = FALSE,
  copy_sty = TRUE,
  line_nums = FALSE,
  line_nums_mod = 1,
  draft_watermark = FALSE,
  highlight = "tango",
  french = FALSE,
  pandoc_args = c("--top-level-division=chapter", "--wrap=none",
    "--default-image-extension=png"),
  ...
)

techreport_pdf(
  latex_engine = "pdflatex",
  copy_sty = TRUE,
  line_nums = FALSE,
  line_nums_mod = 1,
  lot_lof = FALSE,
  draft_watermark = FALSE,
  highlight = "tango",
  french = FALSE,
  pandoc_args = c("--top-level-division=chapter", "--wrap=none",
    "--default-image-extension=png"),
  ...
)
}
\arguments{
\item{toc}{Logical. If \code{TRUE}, include a Table of Contents (TOC)}

\item{toc_depth}{A positive integer representing how many levels of section
headers to show in the TOC}

\item{highlight}{Syntax highlighting style. Supported styles include
"tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". If not in this list of styles, the directory in
which the build process was called will be searched for the given theme
file. You can copy one from the csasdown library install on your machine
and modify as necessary. Find them by looking here:
file.path(.libPaths(), "csasdown", "themes")
Pass \code{NULL} to prevent syntax highlighting (uses 'monochrome' theme) for
slightly different text format but no highlighting}

\item{latex_engine}{LaTeX engine to render with. 'pdflatex' or 'xelatex'}

\item{prepub}{Logical. If \code{TRUE}, this is a pre-publication version.
Currently only implemented for SR}

\item{copy_sty}{Logical. If \code{TRUE}, the '.sty' files will be copied from
the local library location for every render. If \code{FALSE}, the current '.sty'
file will be kept and any edits you have done to it will remain}

\item{line_nums}{Logical. If \code{TRUE}, include line numbering in the document}

\item{line_nums_mod}{Numerical. Which modulo line numbers to label,
e.g. 2 = every second line, 3 = every 3rd line, etc.}

\item{lot_lof}{Logical. If \code{TRUE}, include list of tables and list of
figures in the document. Implemented only for 'resdoc' and 'techreport'}

\item{draft_watermark}{Logical. If \code{TRUE}, show a DRAFT watermark on all
pages of the output document}

\item{french}{French?}

\item{include_section_nums}{Logical. If \code{TRUE}, include the section and
subsection numbers in the body titles. The table of contents will still
show the numbers}

\item{pandoc_args}{Any other arguments to \code{pandoc}}

\item{...}{other arguments to \code{\link[bookdown:pdf_book]{bookdown::pdf_book()}}.}
}
\value{
Output from \code{\link[bookdown:pdf_book]{bookdown::pdf_book()}}, modified by the post-processor
\code{\link[=fix_envs]{fix_envs()}}
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify using the CSAS LaTeX template and cls files.
}
