% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resdoc_pdf}
\alias{resdoc_pdf}
\alias{sr_pdf}
\alias{techreport_pdf}
\title{Creates an R Markdown PDF with CSAS formatting}
\usage{
resdoc_pdf(
  toc = TRUE,
  toc_depth = 3,
  highlight = "default",
  latex_engine = "pdflatex",
  french = FALSE,
  prepub = FALSE,
  copy_sty = TRUE,
  line_nums = FALSE,
  line_nums_mod = 1,
  pandoc_args = c("--top-level-division=chapter", "--wrap=none",
    "--default-image-extension=png"),
  ...
)

sr_pdf(
  latex_engine = "pdflatex",
  french = FALSE,
  prepub = FALSE,
  copy_sty = TRUE,
  ...
)

techreport_pdf(french = FALSE, latex_engine = "pdflatex", copy_sty = TRUE, ...)
  line_nums = FALSE,
  line_nums_mod = 1,
  pandoc_args = c("--top-level-division=chapter", "--wrap=none",
    "--default-image-extension=png"),
  ...
  french = FALSE,
  latex_engine = "pdflatex",
  copy_sty = TRUE,
  line_nums = FALSE,
  line_nums_mod = 1,
  pandoc_args = c("--top-level-division=chapter", "--wrap=none",
    "--default-image-extension=png"),
  ...
)
}
\arguments{
\item{toc}{A Boolean (\code{TRUE} or \code{FALSE}) specifying whether table of contents
should be created.}

\item{toc_depth}{A positive integer.}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso",
"zenburn", and "haddock". Pass \code{NULL} to prevent syntax highlighting.}

\item{latex_engine}{LaTeX engine.}

\item{french}{Logical for French (vs. English).}

\item{prepub}{Logical for whether this is a prepublication version
(currently not implemented for ResDocs)}

\item{copy_sty}{Copy the .sty files every time? Set to \code{FALSE} to "freeze" the
.sty file if you need to edit it.}

\item{line_nums}{Include line numbers in the document? Logical.}

\item{line_nums_mod}{Numerical. Which modulo line numbers to label, 2 = every second line, etc.}

\item{pandoc_args}{Any other arguments to pandoc.}

\item{...}{other arguments to \code{\link[bookdown:pdf_book]{bookdown::pdf_book()}}.}
}
\value{
A modified \code{pdf_document} based on the CSAS LaTeX template.
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify using the CSAS LaTeX template and cls files.
}
\examples{
\dontrun{
output:csasdown::resdoc_pdf
}
}
