% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{csas_table}
\alias{csas_table}
\title{Custom table for csasdown}
\usage{
csas_table(x, format = "pandoc", booktabs = TRUE, linesep = "",
  longtable = TRUE, font_size = 10, landscape = FALSE,
  repeat_header = TRUE, ...)
}
\arguments{
\item{x}{An R object, typically a matrix or data frame.}

\item{format}{As defined by \code{\link[knitr:kable]{knitr::kable()}}.}

\item{booktabs}{As defined by \code{\link[knitr:kable]{knitr::kable()}}.}

\item{linesep}{As defined by \code{\link[knitr:kable]{knitr::kable()}}.}

\item{longtable}{As defined by \code{\link[knitr:kable]{knitr::kable()}}.}

\item{font_size}{Font size in pts.}

\item{landscape}{Make this table in landscape orientation?}

\item{repeat_header}{If landscape, repeat the header on subsequent pages?}

\item{...}{Other arguments to pass to \code{\link[knitr:kable]{knitr::kable()}}.}
}
\description{
This is a custom wrapper for \code{\link[knitr:kable]{knitr::kable()}} with some arguments set so that
the tables work with CSAS formatting in both LaTeX and Word documents.
}
\examples{
csas_table(head(iris))
}
