% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csasdown.R
\docType{package}
\name{csasdown}
\alias{csasdown}
\title{csasdown: A package for creating CSAS Research Document with R Markdown/bookdown}
\description{
Uses the 'bookdown' R package to generate CSAS (Canadian Science Advisory
Secretariat) Research Documents, Science Responses, and Technical Reports in
PDF or Word format. The package is based on Chester Ismay's thesisdown
package and Ben Marwick's huskydown package.
}
