% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft}
\alias{draft}
\title{Create a new CSAS document based on a template}
\usage{
draft(
  type = c("resdoc", "fsar", "sr", "techreport"),
  directory = ".",
  edit = FALSE,
  create_rstudio_project = TRUE,
  ...
)
}
\arguments{
\item{type}{The type of document to draft. Must be one of \code{resdoc}, \code{fsar},
\code{sr}, or \code{techreport}.}

\item{directory}{The directory to place the draft document files.
Current directory by default}

\item{edit}{\code{TRUE} to edit the template immediately.}

\item{create_rstudio_project}{\code{TRUE} to create an RStudio project file.}

\item{...}{Other arguments to pass to \code{\link[rmarkdown:draft]{rmarkdown::draft()}}.}
}
\description{
Create a draft of an R Markdown CSAS document
}
\details{
This is a light wrapper around \code{\link[rmarkdown:draft]{rmarkdown::draft()}}. Consult that
function for further details.
}
\examples{
\dontrun{
csasdown::draft("resdoc")
csasdown::draft("fsar")
csasdown::draft("techreport")
}
}
