% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft}
\alias{draft}
\title{Create a new CSAS document based on a template}
\usage{
draft(type = c("resdoc", "fsar"), directory = getwd(), edit = FALSE, ...)
}
\arguments{
\item{type}{The type of document to draft. Must be one of \code{resdoc} or
\code{fsar}.}

\item{directory}{The directory to place the draft document files.
Current directory by default}

\item{edit}{\code{TRUE} to edit the template immediately.}

\item{...}{Other arguments to pass to \code{\link[rmarkdown:draft]{rmarkdown::draft()}}.}
}
\description{
Create a draft of an R Markdown CSAS document
}
\details{
This is a light wrapper around \code{\link[rmarkdown:draft]{rmarkdown::draft()}}. Consult that
function for further details.
}
\examples{
\dontrun{
csasdown::draft("resdoc")
csasdown::draft("fsar")
}
}
