% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft}
\alias{draft}
\title{Create a new CSAS document based on a template}
\usage{
draft(
  type = c("resdoc", "resdoc-b", "sr", "techreport"),
  directory = getwd(),
  edit = FALSE,
  verbose = TRUE,
  testing = FALSE,
  testing_affirm = FALSE,
  ...
)
}
\arguments{
\item{type}{The type of document to draft}

\item{directory}{The directory to place the draft document files}

\item{edit}{Logical. If \code{TRUE}, edit the template file immediately
(See \code{edit} in \code{\link[rmarkdown:draft]{rmarkdown::draft()}})}

\item{verbose}{Logical. If \code{TRUE}, print messages}

\item{testing}{Logical. If \code{TRUE}, the question about overwriting
files which already exist will be set to \code{yes} automatically to
ensure unit test work correctly}

\item{testing_affirm}{Logical. If \code{TRUE}, set overwrite to \code{TRUE}
for testing if \code{testing} is \code{TRUE}. If \code{TRUE} set overwrite to \code{FALSE}
for testing if \code{testing} is \code{TRUE}. If \code{testing} is \code{FALSE}, this
will be ignored}

\item{...}{Other arguments to pass to \code{\link[rmarkdown:draft]{rmarkdown::draft()}}.}
}
\description{
Create a draft of an R Markdown CSAS document
}
\details{
This is a light wrapper around \code{\link[rmarkdown:draft]{rmarkdown::draft()}}. Consult that
function for further details.
}
\examples{
\dontrun{
csasdown::draft("resdoc")
csasdown::draft("resdoc-b")
csasdown::draft("sr")
csasdown::draft("techreport")
}
}
\seealso{
Other drafting: 
\code{\link{create_rstudio_project_file}()}
}
\concept{drafting}
