% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft}
\alias{draft}
\title{Create a new CSAS document based on a template}
\usage{
draft(
  type = c("resdoc", "resdoc-b", "sr", "techreport", "manureport"),
  directory = getwd(),
  fn = "index.Rmd",
  edit = FALSE,
  create_rstudio_file = TRUE,
  verbose = TRUE,
  testing = FALSE,
  testing_affirm_ovr = FALSE,
  ...
)
}
\arguments{
\item{type}{The type of document to draft. Must be one of \code{resdoc},
\code{resdoc-b}, \code{sr}, \code{techreport}, or \code{manureport}}

\item{directory}{The directory to place the draft document files.
Is the current directory by default}

\item{fn}{Bookdown starting file, \code{index.Rmd} by default}

\item{edit}{Logical. If \code{TRUE}, edit the template file immediately
(See \code{edit} in \code{\link[rmarkdown:draft]{rmarkdown::draft()}})}

\item{create_rstudio_file}{Logical. Create RStudio project file? Set to
\code{FALSE} if in a subdirectory or a folder that already has an RStudio file.}

\item{verbose}{Logical. If \code{TRUE}, print messages}

\item{testing}{Logical. If \code{TRUE}, the question about overwriting
files which already exist will skipped to ensure unit tests work
correctly}

\item{testing_affirm_ovr}{Logical. If \code{TRUE} and \code{testing} is \code{TRUE},
set overwrite files to \code{TRUE}. If \code{FALSE} and \code{testing} is \code{TRUE},
set overwrite to \code{FALSE}. If \code{testing} is \code{FALSE}, this argument is ignored}

\item{...}{Other arguments to pass to \code{\link[rmarkdown:draft]{rmarkdown::draft()}}.}
}
\description{
Create a draft of an R Markdown CSAS document
}
\details{
This is a light wrapper around \code{\link[rmarkdown:draft]{rmarkdown::draft()}}. Consult that
function for further details.
}
\examples{
\dontrun{
csasdown::draft("resdoc")
csasdown::draft("resdoc-b")
csasdown::draft("sr")
csasdown::draft("techreport")
csasdown::draft("manureport")
}
}
\seealso{
Other drafting: 
\code{\link{create_rstudio_project_file}()}
}
\concept{drafting}
