% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fr}
\alias{fr}
\title{Wrapper for \code{\link[=getOption]{getOption()}} with the default set to FALSE}
\usage{
fr()
}
\value{
The french option value. If the french option is \code{NULL},
\code{FALSE} will be returned
}
\description{
Used to retrieve \code{TRUE} or \code{FALSE} for whether
or not to render the document in French
}
