% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fr}
\alias{fr}
\title{Return value of the \code{french} Option in the current environment}
\usage{
fr()
}
\value{
The \code{french} option value. If the \code{french} option is \code{NULL},
\code{FALSE} will be returned
}
\description{
Used to retrieve \code{TRUE} or \code{FALSE} for whether
or not to render the document in French
}
