% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-book-filename.R
\name{get_book_filename}
\alias{get_book_filename}
\title{Get the book filename from _bookdown.yml file}
\usage{
get_book_filename(fn = "_bookdown.yml")
}
\arguments{
\item{fn}{The Bookdown YAML file name. '_bookdown.yml' by default}
}
\value{
A character string of the supplied book_filename
}
\description{
Get the book filename from _bookdown.yml file
}
\details{
This function reads the _bookdown.yml file to extract the book_filename
}
\keyword{internal}
