% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_email}
\alias{get_email}
\title{Insert regional CSAS email address to contact for information}
\usage{
get_email(region = "National Capital Region")
}
\arguments{
\item{region}{Region in which the document is published; character vector
(i.e., Pacific). Default is "National Capital Region."}
}
\value{
Email address as a character vector
}
\description{
Insert regional CSAS email address to contact for information
}
