% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_email}
\alias{get_email}
\title{Insert regional CSAS email address to contact for information}
\usage{
get_email(region, type = "sr")
}
\arguments{
\item{region}{Region in which the document is published. Character vector
(i.e., Pacific)}

\item{type}{Type of document. Currently this is only implemented for Science
Response documents.}
}
\value{
Email address as a character vector
}
\description{
Insert regional CSAS email address to contact for information
}
