% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject-rmd-files.R
\name{inject_rmd_files}
\alias{inject_rmd_files}
\title{Inject verbatim text from Rmd files into Rmd code}
\usage{
inject_rmd_files(rmd_files)
}
\arguments{
\item{rmd_files}{A vector of character strings representing the names
of Rmd files}
}
\value{
Nothing
}
\description{
Inject verbatim text from Rmd files into Rmd code wrapped in
\code{\link[=cat]{cat()}} so that the bilingual features of csasdown can be used properly
but the author can write paragraphs or sections in pure Rmarkdown
}
\keyword{internal}
