% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_highlight_latex.R
\name{parse_pandoc_highlight_theme}
\alias{parse_pandoc_highlight_theme}
\title{Read in pandoc theme file code as a vector of lines of code
and parse it to create a list of theme tokens}
\usage{
parse_pandoc_highlight_theme(json = NULL, num_attrs = 5)
}
\arguments{
\item{json}{JSON code as a vector of lines of code extracted using a system call like this:
pandoc --print-highlight-style breezedark}

\item{num_attrs}{The number of attributes for each style token in the file.
Style tokens are labeled in the file as text-color, italics, bold, etc.}
}
\value{
A list of 1 element each for the general colors (text and background colors) and 1 element
which is a list of lists of named key-value pairs, where each list represents a chunk of attributes
as defined in the grammar above
}
\description{

}
