% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-comments-from-chunks.R
\name{remove_comments_from_chunks}
\alias{remove_comments_from_chunks}
\title{Remove all comment lines from knitr chunks in the Rmd code}
\usage{
remove_comments_from_chunks(rmd)
}
\arguments{
\item{rmd}{A vector of character strings representing the lines in an
Rmarkdown file}
}
\value{
A modified \code{rmd} vector
}
\description{
Remove all comment lines from knitr chunks in the Rmd code
}
