% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-comments-from-chunks.R
\name{remove_comments_from_chunks}
\alias{remove_comments_from_chunks}
\title{Remove all comment lines from knitr chunks in the Rmd code for the files
provided}
\usage{
remove_comments_from_chunks(rmd_files, verbose = FALSE)
}
\arguments{
\item{rmd_files}{A vector of character strings representing the names
of Rmd files}

\item{verbose}{Logical. If \code{TRUE}, print messages}
}
\value{
A list of length the same as the number of files in \code{rmd_files},
Each list element is another list of two vectors, the first being 1 item for
each chunk in the file and representing the number of comments that were
previously removed in each chunk prior to the \code{cat()}, \verb{<<chunk>>}, or
\code{rmd_file()} call. The second is the number of comments for each chunk
after those calls
}
\description{
Remove all comment lines from knitr chunks in the Rmd code for the files
provided
}
\keyword{internal}
