% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Render a resdoc}
\usage{
render(config_file = "_bookdown.yml", verbose = FALSE, ...)
}
\arguments{
\item{config_file}{YAML configuration file.}

\item{verbose}{Verbose?}

\item{...}{Arguments to pass to \code{\link[bookdown:render_book]{bookdown::render_book()}}.}
}
\description{
Render a resdoc
}
