% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsar-word.R
\name{render_sar}
\alias{render_sar}
\title{Render a SAR}
\usage{
render_sar(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link[bookdown:render_book]{bookdown::render_book()}}.}
}
\description{
Render a SAR
}
