% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_resdoc.R
\name{rmd_filenames_from_yaml}
\alias{rmd_filenames_from_yaml}
\title{Fetch vector of the Rmd filenames found in a bookdown YAML file}
\usage{
rmd_filenames_from_yaml(yaml_fn = "_bookdown.yml")
}
\arguments{
\item{yaml_fn}{The YAML file name}
}
\value{
A vector of the Rmd file names
}
\description{
Return a vector of the names of all Rmd files found in the file \code{yaml_fn}.
It is required that all filenames appear on their own line in the file.
Commented-out filenames will not be included.
}
\details{
The file must have an entry like this:

\verb{rmd_files: ["index.Rmd",}
\verb{            "01-chap1.Rmd"]}
\verb{           #"02-chap2.Rmd",}
\verb{           #"03-chap3.Rmd",}
\verb{           #"04-references.Rmd",}
\verb{           #"05-appendix.Rmd"]}

In this case, the vector returned will be:

c("index.Rmd", "01-chap1.Rmd")
}
