% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-render-type.R
\name{set_render_type}
\alias{set_render_type}
\title{Set the the type of document to render by modifying the YAML tag in
the index file}
\usage{
set_render_type(fn = "index.Rmd", doc_type = "pdf")
}
\arguments{
\item{fn}{The name of the YAML file, typically 'index.Rmd' for bookdown}

\item{doc_type}{The type of document to set for rendering. Either 'pdf' or
'word'}
}
\value{
Nothing
}
\description{
Set the the type of document to render by modifying the YAML tag in
the index file
}
\keyword{internal}
