% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_csasstyle}
\alias{update_csasstyle}
\title{Copy the csas-style directory from the local library location to
the current directory, overwriting.}
\usage{
update_csasstyle(copy = TRUE)
}
\arguments{
\item{copy}{Logical. If TRUE, copy and overwrite if the directory already exists.
If FALSE, only copy if the directory does not exist in the current directory}
}
\value{
Nothing
}
\description{
Copy the csas-style directory from the local library location to
the current directory, overwriting.
}
